Split validate, compile, and hosted CLI responsibilities
Amp-Thread-ID: https://ampcode.com/threads/T-01a01309-c853-743c-a0e8-e1e323f8b930
