+++ internal/cli/cli.go
@@ -14,16 +14,11 @@ import (
 	"os/exec"
 	"os/signal"
 	"path/filepath"
-	"regexp"
 	"runtime"
-	"slices"
 	"sort"
 	"strings"
-	"sync"
 	"syscall"
 
-	actionintegration "github.com/buildkite/buildkite-gha/internal/action/integration"
-	actionsource "github.com/buildkite/buildkite-gha/internal/action/source"
 	buildkitepipeline "github.com/buildkite/buildkite-gha/internal/buildkite"
 	"github.com/buildkite/buildkite-gha/internal/compatibility"
 	"github.com/buildkite/buildkite-gha/internal/compiler"
@@ -38,18 +33,10 @@ const (
 	legacyRuntimeQueue            = "hosted"
 	legacyTargetQueueEnvironment  = "BUILDKITE_GHA_TARGET_QUEUE"
 	legacyRuntimeImageEnvironment = "BUILDKITE_GHA_RUNTIME_IMAGE"
-	hostedProfile                 = "hosted"
-	legacyHostedTokenlessProfile  = "hosted-tokenless"
 	maxWebhookMetadataBytes       = 25 << 20
 	workflowCheckSummaryLimit     = 65535
 	workflowCheckSummaryNotice    = "\n\n_Additional diagnostics omitted at the provider check summary size limit._\n"
-	defaultNobleRunnerImage       = "buildkite.namespace-images.com/agent-base@sha256:d8fa20ce298d9ee8673dc6e9ecddabb23ebbff3bcaf828d2011714a960cc5853"
-	defaultJammyRunnerImage       = "buildkite.namespace-images.com/agent-base@sha256:3489d7d999869a11a243bdd9f91f318784cb74f0b57646c57424d10008b1e21d"
-	defaultMacOSRunnerQueue       = "macos-medium"
 )
-
-var runnerQueuePattern = regexp.MustCompile(`^[A-Za-z0-9_-]{1,255}$`)
-var runnerImagePattern = regexp.MustCompile(`^[a-z0-9]+(?:[._-][a-z0-9]+)*(?:/[a-z0-9]+(?:[._-][a-z0-9]+)*)+@sha256:[0-9a-f]{64}$`)
 
 // Run executes the command and returns its process exit code.
 func Run(args []string, stdout, stderr io.Writer, version string) int {
@@ -112,477 +99,6 @@ func commandVersion(clientVersion string) string {
 		return "dev"
 	}
 	return clientVersion
-}
-
-func validate(args []string, stdout, stderr io.Writer, version string, agent transport.Agent) int {
-	args, actionCacheDir, err := validateActionCacheArgs(args)
-	if err != nil {
-		return usageError(stderr, "validate: %v", err)
-	}
-	workflowPath, eventPath, eventName, format, profile, allEvents, err := validateArgs(args)
-	if err != nil {
-		return usageError(stderr, "validate: %v", err)
-	}
-	if actionCacheDir != "" && profile == "" {
-		return usageError(stderr, "validate: --action-cache-dir requires --profile hosted")
-	}
-	if actionCacheDir != "" {
-		if _, err := actionsource.NewStore(actionCacheDir, nil); err != nil {
-			return usageError(stderr, "validate: --action-cache-dir: %v", err)
-		}
-	}
-	if profile != "" && eventPath == "" && eventName == "" && !allEvents {
-		return usageError(stderr, "validate: --profile hosted requires --event, --event-path, or --all-events; use bare validate <workflow> for event-independent syntax and trigger compatibility validation")
-	}
-	out := newProcessingOutput(context.Background(), "validate", format, stdout, stderr, agent)
-	if allEvents {
-		return validateAllEvents(out, workflowPath, version, actionCacheDir, nil, stderr)
-	}
-	return validateOne(out, workflowPath, eventPath, eventName, profile, version, actionCacheDir, nil, stderr)
-}
-
-func validateOne(out processingOutput, workflowPath, eventPath, eventName, profile, version, actionCacheDir string, runtime *profileValidationRuntime, stderr io.Writer) int {
-	return validateOneSource(out, workflowPath, nil, eventPath, eventName, profile, version, actionCacheDir, runtime, stderr)
-}
-
-func validateOneSource(out processingOutput, workflowPath string, workflowSource []byte, eventPath, eventName, profile, version, actionCacheDir string, runtime *profileValidationRuntime, stderr io.Writer) int {
-	contextRequired := false
-	var loadEvent func() ([]byte, error)
-	if eventPath != "" {
-		event, eventErr := os.ReadFile(eventPath)
-		if parsedEvent, parseErr := compiler.ParseEvent(event); eventErr == nil && parseErr == nil {
-			out.sourceLinks = sourceLinksForEvent(parsedEvent)
-		}
-		loadEvent = func() ([]byte, error) { return event, eventErr }
-	} else if eventName != "" {
-		event, eventErr := generatedEventSnapshot(eventName)
-		loadEvent = func() ([]byte, error) { return event, eventErr }
-	}
-	var source, event []byte
-	var ok bool
-	if workflowSource == nil {
-		source, event, ok = loadProcessingInputs(out, workflowPath, profile, "event input could not be read", loadEvent)
-	} else {
-		source, event, ok = loadProcessingInputsSource(out, workflowPath, profile, workflowSource, "event input could not be read", loadEvent)
-	}
-	if !ok {
-		return 1
-	}
-	if profile != "" {
-		parsed, parseErr := workflow.Parse(workflowPath, source)
-		effectiveEvent, eventErr := newEffectiveEvent(event, effectiveEventFromPath, os.Getenv)
-		if parseErr == nil && eventErr == nil && !parsed.ReusableOnly() {
-			selection, triggerErr := selectWorkflowTrigger(parsed.Triggers, effectiveEvent)
-			if triggerErr != nil {
-				contextRequired = pathFilterContextRequired(triggerErr)
-				if contextRequired {
-					triggerErr = buildkitepipeline.ValidateTriggerConditions(parsed.Triggers)
-					contextRequired = triggerErr == nil
-				}
-				if !contextRequired {
-					report := triggerFailureProcessingReport(workflowInput{Path: workflowPath, Source: source}, triggerErr)
-					_ = out.write(report)
-					return 1
-				}
-			} else if !selection.Applicable {
-				report := triggerProcessingReport(workflowPath, source)
-				report.Result = "not-applicable"
-				if out.write(report) != nil {
-					return 1
-				}
-				return 0
-			}
-		}
-	}
-	// Profile validation applies the hosted runner policy so validate and
-	// upload agree on supported labels, including the default macOS queue.
-	var validationOptions *compiler.Options
-	if profile != "" {
-		hosted := hostedOptions("", nil, nil)
-		validationOptions = &hosted
-	}
-	processingReport, ok := validatedProcessingReportWithOptions(out, workflowPath, profile, source, event, loadEvent != nil, validationOptions)
-	if !ok {
-		return 1
-	}
-	if profile != "" {
-		distributionDigest := ""
-		var executableErr error
-		if runtime != nil {
-			distributionDigest, executableErr = runtime.distributionDigest, runtime.executableErr
-		} else {
-			_, _, distributionDigest, executableErr = executable()
-		}
-		if executableErr != nil {
-			processingReport.AddEnvironmentFailure("compiler executable could not be inspected")
-			processingReport.Result = "indeterminate"
-			_ = out.write(processingReport)
-			return 1
-		}
-		ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
-		defer stop()
-		// Profile validation never executes generated plans, so the importer
-		// digest stands in for every platform's runtime distribution.
-		runtimeDistributions := map[compiler.Platform]string{
-			compiler.PlatformLinuxAMD64:  distributionDigest,
-			compiler.PlatformDarwinARM64: distributionDigest,
-		}
-		var actionSource compiler.ActionSource
-		if runtime != nil {
-			actionSource = runtime.actionSource
-		}
-		preflight, profileErr := compileHostedWithActionCache(ctx, workflowPath, source, event, version, distributionDigest, "buildkite-gha-profile-importer", "", nil, runtimeDistributions, actionCacheDir, actionSource, nil)
-		applyHostedPreflight(&processingReport, preflight)
-		if profileErr != nil {
-			if ctx.Err() != nil || errors.Is(profileErr, context.Canceled) {
-				_, _ = fmt.Fprintf(stderr, "buildkite-gha: validate: profile evaluation interrupted: %v\n", profileErr)
-				return 1
-			}
-			processingReport.Result = classifyHostedFailure(&processingReport, workflowPath, profileErr)
-			_ = out.write(processingReport)
-			return 1
-		}
-		if preflight.HasActions {
-			processingReport.Diagnostics = append(processingReport.Diagnostics, compatibility.Diagnostic{
-				Level: "warning", Code: "W_ACTION_RUNTIME_UNKNOWN", Category: "compatibility", Stage: string(compiler.StageAdmission),
-				Message: "Action runtime behavior was not evaluated. The action source, metadata, declared runtime, entrypoints, inputs, and nested actions were validated, but the action code was not executed and may depend on GitHub-only services. No action is required for admission; test the action on Buildkite if runtime compatibility is important.",
-			})
-		}
-		if contextRequired {
-			processingReport.SetStage(string(compiler.StageAdmission), compatibility.NotEvaluated)
-			processingReport.Admission.Result = compatibility.NotEvaluated
-			processingReport.Diagnostics = append(processingReport.Diagnostics, compatibility.Diagnostic{
-				Level: "error", Code: compiler.CodeContextRequired, Category: "context", Stage: string(compiler.StageAdmission),
-				Message: "Push and pull request path filters require linked Buildkite webhook data and a verified local git diff before admission can be determined.",
-			})
-			processingReport.Result = "context-required"
-			if out.write(processingReport) != nil {
-				return 1
-			}
-			return 1
-		}
-		processingReport.SetStage(string(compiler.StageAdmission), compatibility.Passed)
-		processingReport.Admission.Result = "admitted"
-		processingReport.Result = "admitted"
-		if out.write(processingReport) != nil {
-			return 1
-		}
-		return 0
-	}
-	processingReport.Result = "compilable"
-	if out.write(processingReport) != nil {
-		return 1
-	}
-	return 0
-}
-
-type profileValidationRuntime struct {
-	actionSource       compiler.ActionSource
-	distributionDigest string
-	executableErr      error
-}
-
-func validateAllEvents(out processingOutput, workflowPath, version, actionCacheDir string, runtime *profileValidationRuntime, stderr io.Writer) int {
-	source, err := os.ReadFile(workflowPath)
-	if err != nil {
-		validation := compatibility.EnvironmentProcessingReport(workflowPath, "", "workflow input could not be read")
-		report := compatibility.NewProcessingReportV3(workflowPath, hostedProfile, validation)
-		_ = out.writeV3(report)
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: validate: %v\n", err)
-		return 1
-	}
-	return validateAllEventsSource(out, workflowPath, source, version, actionCacheDir, runtime, stderr)
-}
-
-func validateAllEventsSource(out processingOutput, workflowPath string, source []byte, version, actionCacheDir string, runtime *profileValidationRuntime, stderr io.Writer) int {
-	validation, validationErr := compiler.ValidateWithOptions(workflowPath, source, hostedOptions("", nil, nil))
-	validationReport := compatibility.InitialProcessingReport(workflowPath, "", false, validation, validationErr)
-	if validationErr != nil {
-		validationReport.Result = "incompatible"
-		report := compatibility.NewProcessingReportV3(workflowPath, hostedProfile, validationReport)
-		if out.writeV3(report) != nil {
-			return 1
-		}
-		return 1
-	}
-	validationReport.Result = "compilable"
-	report := compatibility.NewProcessingReportV3(workflowPath, hostedProfile, validationReport)
-	parsed, err := workflow.Parse(workflowPath, source)
-	if err != nil {
-		return 1
-	}
-	declared := make(map[string]bool, len(parsed.Triggers))
-	for _, trigger := range parsed.Triggers {
-		declared[trigger.Event] = true
-	}
-	cleanup := func() {}
-	if runtime == nil {
-		actionSource, sourceCleanup, sourceErr := newHostedActionSource(actionCacheDir, nil, nil)
-		cleanup = sourceCleanup
-		if sourceErr != nil {
-			validationReport.AddEnvironmentFailure("public action source could not be configured")
-			validationReport.Result = "indeterminate"
-			report.Validation = validationReport
-			_ = out.writeV3(report)
-			_, _ = fmt.Fprintf(stderr, "buildkite-gha: validate: %v\n", sourceErr)
-			cleanup()
-			return 1
-		}
-		_, _, distributionDigest, executableErr := executable()
-		runtime = &profileValidationRuntime{actionSource: actionSource, distributionDigest: distributionDigest, executableErr: executableErr}
-	}
-	defer cleanup()
-	failed := false
-	for _, event := range []string{"push", "pull_request", "merge_group", "release", "workflow_dispatch", "schedule"} {
-		if !declared[event] {
-			continue
-		}
-		var eventReport *compatibility.ProcessingReport
-		eventOut := processingOutput{
-			command: "validate", format: "json", reports: io.Discard, stderr: stderr,
-			observe: func(observed compatibility.ProcessingReport) { eventReport = &observed },
-		}
-		if code := validateOneSource(eventOut, workflowPath, source, "", event, hostedProfile, version, actionCacheDir, runtime, stderr); code != 0 {
-			failed = true
-		}
-		if eventReport == nil {
-			return 1
-		}
-		report.Evaluations = append(report.Evaluations, compatibility.EventEvaluation{
-			Event: event, Source: "generated", Report: *eventReport,
-		})
-	}
-	if out.writeV3(report) != nil {
-		return 1
-	}
-	if failed {
-		return 1
-	}
-	return 0
-}
-
-func validateActionCacheArgs(args []string) ([]string, string, error) {
-	filtered := make([]string, 0, len(args))
-	var cacheDir string
-	seen := false
-	for i := 0; i < len(args); i++ {
-		if args[i] != "--action-cache-dir" {
-			filtered = append(filtered, args[i])
-			continue
-		}
-		if seen {
-			return nil, "", fmt.Errorf("--action-cache-dir may only be specified once")
-		}
-		seen = true
-		i++
-		if i == len(args) || strings.TrimSpace(args[i]) == "" {
-			return nil, "", fmt.Errorf("--action-cache-dir requires a path")
-		}
-		cacheDir = args[i]
-	}
-	return filtered, cacheDir, nil
-}
-
-func validateArgs(args []string) (workflowPath, eventPath, eventName, format, profile string, allEvents bool, err error) {
-	format = "text"
-	formatSeen := false
-	profileSeen := false
-	eventPathSeen := false
-	eventSeen := false
-	allEventsSeen := false
-	for i := 0; i < len(args); i++ {
-		option := args[i]
-		switch option {
-		case "--format", "--profile", "--event-path", "--event":
-			seen := map[string]bool{"--format": formatSeen, "--profile": profileSeen, "--event-path": eventPathSeen, "--event": eventSeen}[option]
-			if seen {
-				return "", "", "", "", "", false, fmt.Errorf("%s may only be specified once", option)
-			}
-			i++
-			if i == len(args) {
-				return "", "", "", "", "", false, fmt.Errorf("%s requires a value", option)
-			}
-		case "--all-events":
-			if allEventsSeen {
-				return "", "", "", "", "", false, fmt.Errorf("--all-events may only be specified once")
-			}
-			allEventsSeen = true
-			allEvents = true
-			continue
-		case "-h", "--help":
-			return "", "", "", "", "", false, fmt.Errorf("help must be requested immediately after the command")
-		default:
-			if strings.HasPrefix(option, "-") {
-				return "", "", "", "", "", false, fmt.Errorf("unknown option %q", option)
-			}
-			if workflowPath != "" {
-				return "", "", "", "", "", false, fmt.Errorf("expected one workflow path")
-			}
-			workflowPath = option
-			continue
-		}
-		switch option {
-		case "--format":
-			formatSeen = true
-			format = args[i]
-			if format != "text" && format != "json" {
-				return "", "", "", "", "", false, fmt.Errorf("--format must be text or json")
-			}
-		case "--profile":
-			profileSeen = true
-			profile = args[i]
-			if profile != hostedProfile && profile != legacyHostedTokenlessProfile {
-				return "", "", "", "", "", false, fmt.Errorf("--profile must be %q", hostedProfile)
-			}
-			profile = hostedProfile
-		case "--event-path":
-			eventPathSeen = true
-			eventPath = args[i]
-		case "--event":
-			eventSeen = true
-			eventName = args[i]
-		}
-	}
-	if workflowPath == "" {
-		return "", "", "", "", "", false, fmt.Errorf("workflow path is required")
-	}
-	if eventPathSeen && eventSeen {
-		return "", "", "", "", "", false, fmt.Errorf("--event and --event-path are mutually exclusive")
-	}
-	if eventSeen && profile == "" {
-		return "", "", "", "", "", false, fmt.Errorf("--event requires --profile hosted")
-	}
-	if eventSeen && !slices.Contains([]string{"push", "pull_request", "merge_group", "release", "workflow_dispatch", "schedule"}, eventName) {
-		return "", "", "", "", "", false, fmt.Errorf("unsupported --event %q; supported events are push, pull_request, merge_group, release, workflow_dispatch, and schedule", eventName)
-	}
-	if allEvents && (eventPathSeen || eventSeen) {
-		return "", "", "", "", "", false, fmt.Errorf("--all-events is mutually exclusive with --event and --event-path")
-	}
-	if allEvents && profile == "" {
-		return "", "", "", "", "", false, fmt.Errorf("--all-events requires --profile hosted")
-	}
-	return workflowPath, eventPath, eventName, format, profile, allEvents, nil
-}
-
-func generatedEventSnapshot(name string) ([]byte, error) {
-	event := struct {
-		Provider   string              `json:"provider"`
-		Event      string              `json:"event"`
-		Repository compiler.Repository `json:"repository"`
-		Ref        string              `json:"ref"`
-		SHA        string              `json:"sha"`
-		Actor      string              `json:"actor"`
-		Payload    map[string]any      `json:"payload"`
-	}{
-		Provider: "github", Event: name,
-		Repository: compiler.Repository{Owner: "example", Name: "repository", CloneURL: "https://github.com/example/repository.git", DefaultBranch: "main"},
-		Ref:        "refs/heads/main", SHA: strings.Repeat("0", 40), Actor: "buildkite-gha", Payload: map[string]any{},
-	}
-	switch name {
-	case "push":
-		event.Payload["ref"] = event.Ref
-	case "pull_request":
-		event.Ref = "refs/pull/1/merge"
-		event.Payload["action"] = "opened"
-		event.Payload["number"] = 1
-		event.Payload["pull_request"] = map[string]any{
-			"number": 1,
-			"base":   map[string]any{"ref": "main"},
-			"head":   map[string]any{"ref": "example"},
-		}
-	case "merge_group":
-		event.Ref = "refs/heads/gh-readonly-queue/main/pr-1-deadbeef"
-		event.Payload["action"] = "checks_requested"
-		event.Payload["merge_group"] = map[string]any{
-			"head_ref": event.Ref,
-			"head_sha": event.SHA,
-			"base_ref": "refs/heads/main",
-			"base_sha": strings.Repeat("1", 40),
-		}
-	case "release":
-		event.Ref = "refs/tags/v1.0.0"
-		event.Payload["action"] = "published"
-		event.Payload["release"] = map[string]any{
-			"tag_name":   "v1.0.0",
-			"draft":      false,
-			"prerelease": false,
-		}
-	case "schedule":
-		event.Payload["schedule"] = "0 0 * * *"
-	case "workflow_dispatch":
-	default:
-		return nil, fmt.Errorf("unsupported generated event %q; supported events are push, pull_request, merge_group, release, workflow_dispatch, and schedule", name)
-	}
-	return json.Marshal(event)
-}
-
-func compile(args []string, stdout, stderr io.Writer, version string, agent transport.Agent) int {
-	workflowPath, eventPath, format, err := compileArgs(args)
-	if err != nil {
-		return usageError(stderr, "compile: %v", err)
-	}
-	if eventPath == "" {
-		return usageError(stderr, "compile: --event-path is required")
-	}
-	out := newProcessingOutput(context.Background(), "compile", "text", stderr, stderr, agent)
-	event, eventErr := os.ReadFile(eventPath)
-	if parsedEvent, parseErr := compiler.ParseEvent(event); eventErr == nil && parseErr == nil {
-		out.sourceLinks = sourceLinksForEvent(parsedEvent)
-	}
-	source, event, ok := loadProcessingInputs(out, workflowPath, "", "event input could not be read", func() ([]byte, error) { return event, eventErr })
-	if !ok {
-		return 1
-	}
-	processingReport, ok := validatedProcessingReport(out, workflowPath, "", source, event, true)
-	if !ok {
-		return 1
-	}
-	var result []byte
-	var warnings []compiler.Warning
-	if format == "ir-json" {
-		result, err = compiler.Compile(workflowPath, source, event)
-		if err == nil {
-			var ir compiler.IR
-			if decodeErr := json.Unmarshal(result, &ir); decodeErr != nil {
-				err = fmt.Errorf("decode compiler IR: %w", decodeErr)
-			} else {
-				warnings = ir.Warnings
-			}
-		}
-	} else {
-		digest, digestErr := executableDigest()
-		if digestErr != nil {
-			processingReport.AddEnvironmentFailure("compiler executable could not be inspected")
-			processingReport.Result = "indeterminate"
-			_ = out.write(processingReport)
-			return 1
-		}
-		bundle, compileErr := compiler.CompileBundle(workflowPath, source, event, version, digest, "gha-importer")
-		processingReport.ApplyEvidence(bundle.Processing)
-		err = compileErr
-		result = bundle.Pipeline
-		warnings = bundle.IR.Warnings
-	}
-	if err != nil {
-		processingReport.AddFailure(workflowPath, string(compiler.StagePlans), compiler.CodePlanConstruction, "compatibility", err)
-		processingReport.Result = "incompatible"
-		_ = out.write(processingReport)
-		return 1
-	}
-	processingReport.Result = "compilable"
-	_ = out.write(processingReport)
-	writeCompilerWarnings(stderr, "compile", workflowPath, warnings)
-	if _, err := stdout.Write(result); err != nil {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: compile: write output: %v\n", err)
-		return 1
-	}
-	return 0
-}
-
-func writeCompilerWarnings(stderr io.Writer, command, path string, warnings []compiler.Warning) {
-	for _, warning := range warnings {
-		_, _ = fmt.Fprintf(stderr, "buildkite-gha: %s: warning: %s:%d:%d: [%s] %s\n", command, path, warning.Line, warning.Column, warning.Code, warning.Message)
-	}
 }
 
 func upload(args []string, stdout, stderr io.Writer, version string, agent transport.Agent) int {
@@ -1059,18 +575,6 @@ func requiredRuntimePlatforms(workflowPath string, workflowSource, eventSource [
 	return platforms, nil
 }
 
-func configuredRunnerPlatform(labels []string, configuredTargets map[string]compiler.RunnerTarget) (compiler.Platform, error) {
-	if len(labels) == 0 {
-		return compiler.Platform{}, fmt.Errorf("runs-on resolved to no labels")
-	}
-	canonical := strings.ToLower(strings.TrimSpace(labels[0]))
-	if target, ok := configuredTargets[canonical]; ok {
-		return target.Platform, nil
-	}
-	_, platform, err := supportedRunnerTarget(canonical)
-	return platform, err
-}
-
 type effectiveEventOrigin string
 
 const (
@@ -1356,439 +860,6 @@ func workflowInputs(matches []workflowInput, namespaceKeys bool) ([]workflowInpu
 	return out, nil
 }
 
-type hostedCompilation struct {
-	Bundle     compiler.Bundle
-	HasActions bool
-	Admitted   bool
-}
-
-type hostedFailureKind string
-
-const (
-	hostedEnvironmentFailure hostedFailureKind = "environment"
-	hostedEvaluationFailure  hostedFailureKind = "evaluation"
-	hostedAdmissionFailure   hostedFailureKind = "admission"
-)
-
-type hostedFailure struct {
-	Kind hostedFailureKind
-	Err  error
-}
-
-func (e *hostedFailure) Error() string { return e.Err.Error() }
-func (e *hostedFailure) Unwrap() error { return e.Err }
-
-func hostedError(kind hostedFailureKind, err error) error {
-	return &hostedFailure{Kind: kind, Err: err}
-}
-
-type actionSourceAuthentication struct {
-	provider   gharuntime.ActionSourceTokenProvider
-	redactor   gharuntime.Redactor
-	warnings   io.Writer
-	once       sync.Once
-	credential string
-	err        error
-}
-
-func importerJobActionSourceAuthentication(warnings io.Writer) *actionSourceAuthentication {
-	authentication := &actionSourceAuthentication{
-		redactor: gharuntime.AgentRedactor{Executable: os.Getenv("BUILDKITE_GHA_AGENT")},
-		warnings: warnings,
-	}
-	provider, err := gharuntime.NewAgentGitHubTokens(gharuntime.AgentGitHubTokenConfig{
-		Endpoint: os.Getenv("BUILDKITE_AGENT_ENDPOINT"),
-		JobID:    os.Getenv("BUILDKITE_JOB_ID"),
-		JobToken: os.Getenv("BUILDKITE_AGENT_ACCESS_TOKEN"),
-	})
-	if err != nil {
-		return authentication
-	}
-	authentication.provider = provider
-	return authentication
-}
-
-func (a *actionSourceAuthentication) option(repository string) actionsource.Option {
-	if a == nil {
-		return nil
-	}
-	return actionsource.WithGitHubActionSourceTokenProvider(repository, func(ctx context.Context) (string, error) {
-		return a.token(ctx, repository)
-	})
-}
-
-func (a *actionSourceAuthentication) token(ctx context.Context, repository string) (string, error) {
-	if err := ctx.Err(); err != nil {
-		return "", err
-	}
-	a.once.Do(func() {
-		a.credential, a.err = a.provision(ctx, repository)
-	})
-	return a.credential, a.err
-}
-
-func (a *actionSourceAuthentication) provision(ctx context.Context, repository string) (string, error) {
-	if a.provider == nil {
-		a.warnAnonymousFallback("GitHub action source authentication is unavailable")
-		return "", nil
-	}
-	token, err := a.provider.ActionSourceToken(ctx, repository)
-	if err != nil {
-		if ctx.Err() != nil {
-			return "", ctx.Err()
-		}
-		if errors.Is(err, context.Canceled) {
-			return "", err
-		}
-		a.warnAnonymousFallback("could not mint a GitHub action source token")
-		return "", nil
-	}
-	if err := a.redactor.AddRedaction(ctx, token); err != nil {
-		if ctx.Err() != nil {
-			return "", ctx.Err()
-		}
-		if errors.Is(err, context.Canceled) {
-			return "", err
-		}
-		a.warnAnonymousFallback("could not register the GitHub action source token with the Buildkite Agent redactor")
-		return "", nil
-	}
-	return token, nil
-}
-
-func (a *actionSourceAuthentication) warnAnonymousFallback(reason string) {
-	if a.warnings != nil {
-		_, _ = fmt.Fprintf(a.warnings, "buildkite-gha: upload: warning: %s; resolving mutable public action references anonymously\n", reason)
-	}
-}
-
-func hostedOptions(groupLabel string, configuredTargets map[string]compiler.RunnerTarget, runtimeDistributions map[compiler.Platform]string) compiler.Options {
-	targets := hostedRunnerTargets()
-	for label, target := range configuredTargets {
-		targets[label] = target
-	}
-	options := compiler.Options{
-		EventTrust:           compiler.EventUntrusted,
-		GroupLabel:           groupLabel,
-		RuntimeDistributions: runtimeDistributions,
-		Runners: compiler.RunnerPolicy{
-			Targets:                    targets,
-			AllowUntrustedDefaultQueue: true,
-		},
-	}
-	for _, target := range targets {
-		if target.Queue != "" && !slices.Contains(options.Runners.UntrustedQueues, target.Queue) {
-			options.Runners.UntrustedQueues = append(options.Runners.UntrustedQueues, target.Queue)
-		}
-	}
-	return options
-}
-
-// hostedRunnerTargets is the runner preset shared by hosted validation and
-// production upload. RunnerPolicy resolves these labels case-insensitively.
-// Keep version-specific and organization-provided targets out of this preset.
-func hostedRunnerTargets() map[string]compiler.RunnerTarget {
-	return map[string]compiler.RunnerTarget{
-		"ubuntu-latest": {Platform: compiler.PlatformLinuxAMD64, Image: defaultNobleRunnerImage},
-		"ubuntu-24.04":  {Platform: compiler.PlatformLinuxAMD64, Image: defaultNobleRunnerImage},
-		"ubuntu-22.04":  {Platform: compiler.PlatformLinuxAMD64, Image: defaultJammyRunnerImage},
-		"macos-latest":  {Platform: compiler.PlatformDarwinARM64, Queue: defaultMacOSRunnerQueue},
-	}
-}
-
-func compileHosted(ctx context.Context, workflowPath string, workflowSource, eventSource []byte, version, distributionDigest, importerStep, groupLabel string, configuredTargets map[string]compiler.RunnerTarget, runtimeDistributions map[compiler.Platform]string, actionAuthentication *actionSourceAuthentication) (hostedCompilation, error) {
-	return compileHostedNamespaced(ctx, workflowPath, workflowSource, eventSource, version, distributionDigest, importerStep, groupLabel, configuredTargets, runtimeDistributions, "", nil, actionAuthentication)
-}
-
-func compileHostedWithActionCache(ctx context.Context, workflowPath string, workflowSource, eventSource []byte, version, distributionDigest, importerStep, groupLabel string, configuredTargets map[string]compiler.RunnerTarget, runtimeDistributions map[compiler.Platform]string, actionCacheDir string, sharedActionSource compiler.ActionSource, actionAuthentication *actionSourceAuthentication) (hostedCompilation, error) {
-	return compileHostedNamespacedWithActionCache(ctx, workflowPath, workflowSource, eventSource, version, distributionDigest, importerStep, groupLabel, configuredTargets, runtimeDistributions, "", nil, actionCacheDir, sharedActionSource, actionAuthentication)
-}
-
-func compileHostedNamespaced(ctx context.Context, workflowPath string, workflowSource, eventSource []byte, version, distributionDigest, importerStep, groupLabel string, configuredTargets map[string]compiler.RunnerTarget, runtimeDistributions map[compiler.Platform]string, stepKeyNamespace string, oidc *plan.OIDCConfiguration, actionAuthentication *actionSourceAuthentication) (hostedCompilation, error) {
-	return compileHostedNamespacedWithActionCache(ctx, workflowPath, workflowSource, eventSource, version, distributionDigest, importerStep, groupLabel, configuredTargets, runtimeDistributions, stepKeyNamespace, oidc, "", nil, actionAuthentication)
-}
-
-func compileHostedNamespacedWithActionCache(ctx context.Context, workflowPath string, workflowSource, eventSource []byte, version, distributionDigest, importerStep, groupLabel string, configuredTargets map[string]compiler.RunnerTarget, runtimeDistributions map[compiler.Platform]string, stepKeyNamespace string, oidc *plan.OIDCConfiguration, actionCacheDir string, sharedActionSource compiler.ActionSource, actionAuthentication *actionSourceAuthentication) (hostedCompilation, error) {
-	options := hostedOptions(groupLabel, configuredTargets, runtimeDistributions)
-	options.StepKeyNamespace = stepKeyNamespace
-	options.OIDC = oidc
-	preflight, err := compiler.CompileWithOptions(workflowPath, workflowSource, eventSource, options)
-	if err != nil {
-		return hostedCompilation{}, hostedError(hostedEvaluationFailure, err)
-	}
-	var ir compiler.IR
-	if err := json.Unmarshal(preflight, &ir); err != nil {
-		return hostedCompilation{}, hostedError(hostedEvaluationFailure, fmt.Errorf("decode compiler preflight: %w", err))
-	}
-	hasActions := irUsesActions(ir)
-	if hasActions {
-		actionSource := sharedActionSource
-		cleanup := func() {}
-		if actionSource == nil {
-			var sourceOptions []actionsource.Option
-			if ir.Event.Provider == "github" {
-				authenticationOption := actionAuthentication.option(ir.Event.Repository.Owner + "/" + ir.Event.Repository.Name)
-				if authenticationOption != nil {
-					sourceOptions = append(sourceOptions, authenticationOption)
-				}
-			}
-			actionSource, cleanup, err = newHostedActionSource(actionCacheDir, sourceOptions, nil)
-			if err != nil {
-				return hostedCompilation{}, hostedError(hostedEnvironmentFailure, err)
-			}
-		}
-		defer cleanup()
-		options.ResolveActions = true
-		options.ActionSource = actionSource
-	}
-	bundle, err := compiler.CompileBundlePlansContext(ctx, workflowPath, workflowSource, eventSource, version, distributionDigest, options)
-	if err != nil {
-		return hostedCompilation{Bundle: bundle, HasActions: hasActions}, hostedError(hostedEvaluationFailure, err)
-	}
-	if err := validateUnprivilegedBundle(bundle); err != nil {
-		return hostedCompilation{Bundle: bundle, HasActions: hasActions}, hostedError(hostedAdmissionFailure, err)
-	}
-	bundle, err = compiler.GenerateBundlePipeline(bundle, distributionDigest, importerStep, options)
-	if err != nil {
-		return hostedCompilation{Bundle: bundle, HasActions: hasActions, Admitted: true}, hostedError(hostedEvaluationFailure, err)
-	}
-	if !hasActions && bundleUsesActions(bundle) {
-		return hostedCompilation{Bundle: bundle, HasActions: hasActions, Admitted: true}, hostedError(hostedEvaluationFailure, fmt.Errorf("final compilation introduced actions absent from preflight"))
-	}
-	return hostedCompilation{Bundle: bundle, HasActions: hasActions, Admitted: true}, nil
-}
-
-func newHostedActionSource(actionCacheDir string, resolverOptions, storeOptions []actionsource.Option) (compiler.ActionSource, func(), error) {
-	actionSource, cleanup, _, err := newHostedActionSourceWithSnapshot(actionCacheDir, resolverOptions, storeOptions)
-	return actionSource, cleanup, err
-}
-
-func newHostedActionSourceWithSnapshot(actionCacheDir string, resolverOptions, storeOptions []actionsource.Option) (compiler.ActionSource, func(), string, error) {
-	actionRoot := actionCacheDir
-	cleanup := func() {}
-	if actionRoot == "" {
-		var err error
-		actionRoot, err = os.MkdirTemp("", "buildkite-gha-action-source-")
-		if err != nil {
-			return nil, cleanup, "", fmt.Errorf("create action source store: %w", err)
-		}
-		cleanup = func() { _ = os.RemoveAll(actionRoot) }
-	}
-	resolver, err := actionsource.NewResolver(nil, resolverOptions...)
-	if err != nil {
-		cleanup()
-		return nil, func() {}, "", fmt.Errorf("configure public action resolver: %w", err)
-	}
-	store, err := actionsource.NewStore(actionRoot, nil, storeOptions...)
-	if err != nil {
-		cleanup()
-		return nil, func() {}, "", fmt.Errorf("configure public action source store: %w", err)
-	}
-	return compiler.MemoizeActionSource(compiler.PublicActionSource{Resolver: resolver, Store: store}), cleanup, resolver.ResolutionSnapshotID(), nil
-}
-
-func validateUnprivilegedBundle(bundle compiler.Bundle) error {
-	instances := make(map[string]compiler.JobInstance, len(bundle.IR.Jobs))
-	for _, instance := range bundle.IR.Jobs {
-		instances[instance.Key] = instance
-	}
-	var diagnostics []error
-	addFailure := func(artifact compiler.PlanArtifact, message, detail string, err error) {
-		finding := &compiler.ProcessingFinding{
-			Stage: compiler.StageAdmission, Code: "E_PROFILE", Category: "admission",
-			Job: artifact.Job.Workflow.LogicalJobID, Instance: artifact.Job.Target.StepKey,
-			Message: message, Detail: detail, Err: err,
-		}
-		if instance, ok := instances[artifact.Job.Target.StepKey]; ok {
-			finding.Path = instance.SourcePath
-			finding.Line = instance.Source.Start.Line
-			finding.Column = instance.Source.Start.Column
-		}
-		diagnostics = append(diagnostics, finding)
-	}
-	for _, artifact := range bundle.Plans {
-		for _, capability := range artifact.Job.RequiredCapabilities {
-			if capability == "secrets" {
-				continue
-			}
-			if capability == "docker" && !admittedDockerProvenance(artifact.Job, artifact.Authorization.DockerCapabilitySources) {
-				message := fmt.Sprintf("Job %q requires Docker without matching compiler provenance. Hosted runs support only verified Dockerfile actions and bounded job or service containers.", artifact.Job.Workflow.LogicalJobID)
-				addFailure(artifact, message, "", errors.New("unsupported Docker access"))
-				continue
-			}
-			if capability == "provider-token-read" {
-				if !slices.Equal(artifact.Authorization.ProviderTokenReadCapabilitySources, []string{"checkout-adapter"}) {
-					message := "GitHub checkout credentials are unsupported for this job; use the managed actions/checkout integration"
-					addFailure(artifact, message, "", errors.New("unsupported GitHub checkout credentials"))
-				}
-				continue
-			}
-			if capability == "provider-token-write" {
-				filename := ""
-				if artifact.Job.GitHubToken != nil {
-					filename = artifact.Job.GitHubToken.Workflow
-				}
-				filenameErr := plan.ValidateGitHubWorkflowPolicyFilename(filename)
-				if artifact.Job.GitHubToken == nil || !slices.Equal(artifact.Authorization.ProviderTokenWriteCapabilitySources, []string{"effective-permissions"}) || filenameErr != nil || artifact.Authorization.WorkflowTokenPolicyFilename == "" || artifact.Authorization.WorkflowTokenPolicyFilename != filename {
-					reason := bundle.IR.Workflow.WorkflowTokenPolicyDiagnostic
-					if reason == "" {
-						reason = "the workflow token request does not match the compiled job"
-					}
-					message, detail := githubTokenAdmissionDiagnostic(artifact, reason)
-					addFailure(artifact, message, detail, errors.New("hosted GITHUB_TOKEN unsupported"))
-				}
-				continue
-			}
-			if capability != "network" && capability != "docker" {
-				message := fmt.Sprintf("Job %q requires unsupported hosted runtime capability %q. Remove the requirement or use a runtime profile that supports it.", artifact.Job.Workflow.LogicalJobID, capability)
-				addFailure(artifact, message, "", errors.New(message))
-			}
-		}
-		for _, action := range artifact.Job.Actions {
-			identity := actionintegration.Identity{Source: action.Source, Repository: action.Repository, Path: action.Path}
-			descriptor, _ := actionintegration.Lookup(identity)
-			if descriptor.Service == actionintegration.ServiceCache {
-				if _, _, err := actionintegration.Admit(identity, action.Commit); err != nil {
-					reference := action.Repository
-					if action.Path != "" {
-						reference += "/" + action.Path
-					}
-					if action.RequestedRef != "" {
-						reference += "@" + action.RequestedRef
-					}
-					message, detail, _ := actionintegration.UnsupportedVersionDiagnostic(reference, err)
-					addFailure(artifact, message, detail, fmt.Errorf("job %q uses unsupported cache action: %w", artifact.Job.Workflow.LogicalJobID, err))
-				}
-				continue
-			}
-			if descriptor.Service != "" {
-				reference := action.Repository
-				if action.Path != "" {
-					reference += "/" + action.Path
-				}
-				message := fmt.Sprintf("Action %q requires the GitHub Actions %s service, which hosted runs do not provide. Replace the action or use a runtime profile that provides this service.", reference, descriptor.Service)
-				addFailure(artifact, message, "", fmt.Errorf("job %q uses action %q, which requires the unavailable GitHub Actions %s service", artifact.Job.Workflow.LogicalJobID, reference, descriptor.Service))
-			}
-		}
-	}
-	return errors.Join(diagnostics...)
-}
-
-func admittedDockerProvenance(job plan.Job, sources []string) bool {
-	if len(sources) == 0 {
-		return false
-	}
-	expected := make([]string, 0, 3)
-	if slices.Contains(sources, "dockerfile-actions") {
-		expected = append(expected, "dockerfile-actions")
-	}
-	if job.Container != nil {
-		expected = append(expected, "job-containers")
-	}
-	if len(job.Services) != 0 || job.ServicesExpression != "" {
-		expected = append(expected, "service-containers")
-	}
-	return slices.Equal(sources, expected)
-}
-
-func githubTokenAdmissionDiagnostic(artifact compiler.PlanArtifact, reason string) (message, detail string) {
-	limitation := reason
-	fix := "Use a supported workflow-level permissions map or remove the GITHUB_TOKEN dependency."
-	switch {
-	case strings.Contains(reason, "job-level permissions"):
-		limitation = "job-level permissions are unsupported for hosted GITHUB_TOKEN issuance"
-		fix = "Move the permissions map to the workflow top level."
-	case strings.Contains(reason, "reusable-workflow jobs"):
-		limitation = "workflows containing reusable-workflow jobs cannot receive a hosted GITHUB_TOKEN"
-		fix = "Inline the reusable job or remove the GITHUB_TOKEN dependency; runner configuration cannot enable this shape."
-	case strings.Contains(reason, "directly under .github/workflows"):
-		limitation = "hosted GITHUB_TOKEN issuance requires the workflow directly under .github/workflows"
-		fix = "Move the workflow directly under .github/workflows."
-	case strings.Contains(reason, "unsupported permission"):
-		limitation = reason
-		fix = "Remove the unsupported permission or avoid requiring GITHUB_TOKEN."
-	case strings.Contains(reason, "explicit non-empty"):
-		limitation = "an explicit empty or all-none top-level permissions map cannot issue GITHUB_TOKEN"
-		fix = "Declare the required permissions at the workflow top level."
-	}
-
-	var causes []string
-	if artifact.Authorization.GitHubTokenSecretReference {
-		causes = append(causes, "the workflow references secrets.GITHUB_TOKEN")
-	}
-	if len(artifact.Authorization.GitHubTokenActions) != 0 {
-		quoted := make([]string, len(artifact.Authorization.GitHubTokenActions))
-		for i, action := range artifact.Authorization.GitHubTokenActions {
-			quoted[i] = fmt.Sprintf("%q", action)
-		}
-		actionLabel := "action"
-		if len(quoted) > 1 {
-			actionLabel = "actions"
-		}
-		causes = append(causes, actionLabel+" "+strings.Join(quoted, ", ")+" defaults an input to github.token")
-	}
-	if len(causes) == 0 {
-		causes = append(causes, "the compiled job requests a workflow token")
-	}
-	permissions := "none"
-	if artifact.Job.GitHubToken != nil && len(artifact.Job.GitHubToken.Permissions) != 0 {
-		names := make([]string, 0, len(artifact.Job.GitHubToken.Permissions))
-		for name := range artifact.Job.GitHubToken.Permissions {
-			names = append(names, name)
-		}
-		sort.Strings(names)
-		values := make([]string, 0, len(names))
-		for _, name := range names {
-			values = append(values, strings.ReplaceAll(name, "_", "-")+": "+artifact.Job.GitHubToken.Permissions[name])
-		}
-		permissions = strings.Join(values, ", ")
-	}
-	message = fmt.Sprintf("Job %q needs GITHUB_TOKEN, but %s. %s", artifact.Job.Workflow.LogicalJobID, limitation, fix)
-	detail = fmt.Sprintf("Cause: %s. Effective permissions: %s.", strings.Join(causes, "; "), permissions)
-	return message, detail
-}
-
-func irUsesActions(ir compiler.IR) bool {
-	for _, job := range ir.Jobs {
-		for _, step := range job.Steps {
-			if step.Uses != "" {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// bundleRunsUnprovenActions reports whether a compiled bundle executes action
-// code the compiler never ran. Actions a native adapter replaces keep both
-// their locks and their `uses` steps in the plan, so the adapter decision is
-// the only thing separating them from actions that really run.
-func bundleRunsUnprovenActions(bundle compiler.Bundle) bool {
-	for _, artifact := range bundle.Plans {
-		for _, lock := range artifact.Job.Actions {
-			identity := actionintegration.Identity{Source: lock.Source, Repository: lock.Repository, Path: lock.Path}
-			if !actionintegration.UsesNativeAdapter(identity) {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-func bundleUsesActions(bundle compiler.Bundle) bool {
-	for _, artifact := range bundle.Plans {
-		if len(artifact.Job.Actions) != 0 {
-			return true
-		}
-		for _, step := range artifact.Job.Steps {
-			if step.Uses != "" || step.Action != nil || step.Kind == "uses" {
-				return true
-			}
-		}
-	}
-	return false
-}
-
 type parsedUploadArgs struct {
 	workflowOperands         []string
 	explicitWorkflowPaths    bool
@@ -1949,101 +1020,4 @@ func parseUploadArgs(args []string) (parsedUploadArgs, error) {
 		}
 	}
 	return parsedUploadArgs{workflowOperands: workflowOperands, eventPath: eventPath, runtimeDistributionPaths: runtimeDistributionPaths, runnerTargets: runnerTargets, experimentalRunnerUser: experimentalRunnerUser}, nil
-}
-
-func configuredRunnerTarget(label, queue, image string) (string, compiler.RunnerTarget, error) {
-	canonical, platform, err := supportedRunnerTarget(label)
-	if err != nil {
-		return "", compiler.RunnerTarget{}, err
-	}
-	if !runnerQueuePattern.MatchString(queue) {
-		return "", compiler.RunnerTarget{}, fmt.Errorf("runner queue for %q is invalid", canonical)
-	}
-	if image != "" && !runnerImagePattern.MatchString(image) {
-		return "", compiler.RunnerTarget{}, fmt.Errorf("runner image for %q must be an immutable registry sha256 reference", canonical)
-	}
-	if image != "" && platform == compiler.PlatformDarwinARM64 {
-		return "", compiler.RunnerTarget{}, fmt.Errorf("runner image for %q is unsupported on darwin/arm64", canonical)
-	}
-	if image == "" {
-		if preset, ok := hostedRunnerTargets()[canonical]; ok {
-			image = preset.Image
-		}
-	}
-	return canonical, compiler.RunnerTarget{Queue: queue, Platform: platform, Image: image}, nil
-}
-
-func supportedRunnerTarget(label string) (string, compiler.Platform, error) {
-	if label != strings.TrimSpace(label) || strings.ContainsAny(label, "\r\n") {
-		return "", compiler.Platform{}, fmt.Errorf("unsupported runner label %q", label)
-	}
-	canonical := strings.ToLower(label)
-	if preset, ok := hostedRunnerTargets()[canonical]; ok {
-		return canonical, preset.Platform, nil
-	}
-	switch canonical {
-	case "macos-15", "macos-14":
-		// Version-specific macOS labels require an organization-provided queue.
-		return canonical, compiler.PlatformDarwinARM64, nil
-	default:
-		return "", compiler.Platform{}, fmt.Errorf("unsupported runner label %q", label)
-	}
-}
-
-func compileArgs(args []string) (workflowPath, eventPath, format string, err error) {
-	format = "pipeline"
-	filtered := make([]string, 0, len(args))
-	formatSeen := false
-	for i := 0; i < len(args); i++ {
-		if args[i] != "--format" {
-			filtered = append(filtered, args[i])
-			continue
-		}
-		if formatSeen {
-			return "", "", "", fmt.Errorf("--format may only be specified once")
-		}
-		formatSeen = true
-		i++
-		if i == len(args) {
-			return "", "", "", fmt.Errorf("--format requires pipeline or ir-json")
-		}
-		format = args[i]
-		if format != "pipeline" && format != "ir-json" {
-			return "", "", "", fmt.Errorf("--format must be pipeline or ir-json")
-		}
-	}
-	workflowPath, eventPath, err = workflowArgs(filtered)
-	return workflowPath, eventPath, format, err
-}
-
-func workflowArgs(args []string) (workflowPath, eventPath string, err error) {
-	eventPathSeen := false
-	for i := 0; i < len(args); i++ {
-		switch args[i] {
-		case "--event-path":
-			if eventPathSeen {
-				return "", "", fmt.Errorf("--event-path may only be specified once")
-			}
-			eventPathSeen = true
-			i++
-			if i == len(args) {
-				return "", "", fmt.Errorf("--event-path requires a path")
-			}
-			eventPath = args[i]
-		case "-h", "--help":
-			return "", "", fmt.Errorf("help must be requested immediately after the command")
-		default:
-			if strings.HasPrefix(args[i], "-") {
-				return "", "", fmt.Errorf("unknown option %q", args[i])
-			}
-			if workflowPath != "" {
-				return "", "", fmt.Errorf("expected one workflow path")
-			}
-			workflowPath = args[i]
-		}
-	}
-	if workflowPath == "" {
-		return "", "", fmt.Errorf("workflow path is required")
-	}
-	return workflowPath, eventPath, nil
 }
