+++ internal/cli/cli_test.go
@@ -7,8 +7,6 @@ import (
 	"errors"
 	"fmt"
 	"io"
-	"net/http"
-	"net/http/httptest"
 	"os"
 	"os/exec"
 	"path/filepath"
@@ -20,8 +18,6 @@ import (
 	"testing"
 	"unicode/utf8"
 
-	actionintegration "github.com/buildkite/buildkite-gha/internal/action/integration"
-	actionsource "github.com/buildkite/buildkite-gha/internal/action/source"
 	buildkitepipeline "github.com/buildkite/buildkite-gha/internal/buildkite"
 	"github.com/buildkite/buildkite-gha/internal/compatibility"
 	"github.com/buildkite/buildkite-gha/internal/compiler"
@@ -66,62 +62,6 @@ func canonicalTempDir(t *testing.T) string {
 		t.Fatal(err)
 	}
 	return dir
-}
-
-func TestConfiguredLinuxRunnerTargetsDefaultHostedToolchainImages(t *testing.T) {
-	for _, test := range []struct {
-		label string
-		image string
-	}{
-		{label: "Ubuntu-Latest", image: defaultNobleRunnerImage},
-		{label: "ubuntu-24.04", image: defaultNobleRunnerImage},
-		{label: "ubuntu-22.04", image: defaultJammyRunnerImage},
-	} {
-		t.Run(test.label, func(t *testing.T) {
-			canonical, target, err := configuredRunnerTarget(test.label, "hosted", "")
-			if err != nil {
-				t.Fatal(err)
-			}
-			if canonical != strings.ToLower(test.label) || target != (compiler.RunnerTarget{Queue: "hosted", Platform: compiler.PlatformLinuxAMD64, Image: test.image}) {
-				t.Fatalf("configuredRunnerTarget() = %q, %#v", canonical, target)
-			}
-		})
-	}
-
-	override := "registry.example.com/custom@sha256:" + strings.Repeat("0", 64)
-	_, target, err := configuredRunnerTarget("ubuntu-latest", "hosted", override)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if target.Image != override {
-		t.Fatalf("explicit image = %q, want %q", target.Image, override)
-	}
-}
-
-func TestHostedRunnerTargetsContainOnlyHostedGuarantees(t *testing.T) {
-	targets := hostedRunnerTargets()
-	labels := make([]string, 0, len(targets))
-	for label := range targets {
-		labels = append(labels, label)
-	}
-	slices.Sort(labels)
-	want := []string{"macos-latest", "ubuntu-22.04", "ubuntu-24.04", "ubuntu-latest"}
-	if !slices.Equal(labels, want) {
-		t.Fatalf("hosted runner labels = %q, want %q", labels, want)
-	}
-	if got := targets["macos-latest"]; got != (compiler.RunnerTarget{Queue: defaultMacOSRunnerQueue, Platform: compiler.PlatformDarwinARM64}) {
-		t.Fatalf("macos-latest target = %#v", got)
-	}
-	for _, label := range []string{"macos-14", "macos-15", "ubuntu-24.04-arm", "windows-latest"} {
-		if _, ok := targets[label]; ok {
-			t.Errorf("hosted runner preset unexpectedly contains %q", label)
-		}
-	}
-
-	canonical, target, err := configuredRunnerTarget("macOS-15", "organization-macos", "")
-	if err != nil || canonical != "macos-15" || target != (compiler.RunnerTarget{Queue: "organization-macos", Platform: compiler.PlatformDarwinARM64}) {
-		t.Fatalf("organization macOS target = %q, %#v, %v", canonical, target, err)
-	}
 }
 
 func TestRunValidateAndCompile(t *testing.T) {
@@ -963,36 +903,6 @@ jobs:
 	}
 }
 
-func TestCompileReportsActionResolutionFailureWithoutPipeline(t *testing.T) {
-	workflowPath := filepath.Join(t.TempDir(), "action.yml")
-	workflow := []byte("on: push\njobs:\n  first:\n    runs-on: ubuntu-latest\n    steps:\n      - uses: ./missing-one\n  second:\n    runs-on: ubuntu-latest\n    steps:\n      - uses: ./missing-two\n")
-	if err := os.WriteFile(workflowPath, workflow, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"compile", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 1 {
-		t.Fatalf("Run() code = %d, want 1; stderr = %q", code, stderr.String())
-	}
-	if stdout.Len() != 0 {
-		t.Fatalf("compile emitted partial pipeline: %q", stdout.String())
-	}
-	for _, want := range []string{
-		"Immutable action resolution: failed",
-		"Job-plan construction: not-evaluated",
-		"Pipeline generation: not-evaluated",
-		"[E_ACTION_RESOLUTION]",
-		"job first/gha-first: passed",
-		"job second/gha-second: passed",
-		"action ./missing-one (job gha-first, step 1): failed",
-		"action ./missing-two (job gha-second, step 1): failed",
-	} {
-		if !strings.Contains(stderr.String(), want) {
-			t.Fatalf("stderr = %q, want %q", stderr.String(), want)
-		}
-	}
-}
-
 func TestProcessingReportStageAttributionDoesNotDependOnWorkflowPath(t *testing.T) {
 	for _, name := range []string{"matrix.yml", "workflow.yml"} {
 		t.Run(name, func(t *testing.T) {
@@ -1017,169 +927,6 @@ func TestProcessingReportStageAttributionDoesNotDependOnWorkflowPath(t *testing.
 				t.Fatalf("report = %#v", report)
 			}
 		})
-	}
-}
-
-func TestCompileLeavesSkippedRemoteActionResolutionNotEvaluated(t *testing.T) {
-	workflowPath := filepath.Join(t.TempDir(), "remote.yml")
-	workflow := []byte(`on: push
-permissions: {}
-jobs:
-  test:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: owner/action@v1
-      - run: echo '${{ secrets.GITHUB_TOKEN }}'
-`)
-	if err := os.WriteFile(workflowPath, workflow, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"compile", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 1 {
-		t.Fatalf("Run() code = %d, want 1; stderr = %q", code, stderr.String())
-	}
-	if stdout.Len() != 0 {
-		t.Fatalf("compile emitted partial pipeline: %q", stdout.String())
-	}
-	for _, want := range []string{
-		"Immutable action resolution: not-evaluated",
-		"Job-plan construction: failed",
-		"action owner/action@v1 (job gha-test, step 1): not-evaluated",
-		"[E_PLAN_CONSTRUCTION]",
-	} {
-		if !strings.Contains(stderr.String(), want) {
-			t.Fatalf("stderr = %q, want %q", stderr.String(), want)
-		}
-	}
-}
-
-func TestHostedReportRetainsIndependentActionInvocationResultsThroughWrapper(t *testing.T) {
-	root := t.TempDir()
-	workflowPath := filepath.Join(root, ".github", "workflows", "actions.yml")
-	actionPath := filepath.Join(root, ".github", "actions", "dynamic")
-	if err := os.MkdirAll(filepath.Dir(workflowPath), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.MkdirAll(actionPath, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	workflow := []byte(`on: push
-jobs:
-  duplicate:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: ./.github/actions/dynamic
-        with:
-          value: supplied
-      - uses: ./.github/actions/dynamic
-  missing:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: ./.github/actions/missing-one
-      - uses: ./.github/actions/missing-two
-`)
-	if err := os.WriteFile(workflowPath, workflow, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	action := []byte(`name: dynamic
-inputs:
-  value:
-    default: ${{ github[env.NAME] }}
-runs:
-  using: node24
-  main: main.js
-`)
-	if err := os.WriteFile(filepath.Join(actionPath, "action.yml"), action, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(actionPath, "main.js"), []byte("console.log('unused')\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	args := []string{"validate", "--profile", "hosted-tokenless", "--format", "json", "--event-path", eventPath, workflowPath}
-	if code := Run(args, &stdout, &stderr, "dev"); code != 1 {
-		t.Fatalf("Run() code = %d, want 1; stderr = %q", code, stderr.String())
-	}
-	var report compatibility.ProcessingReport
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	if len(report.Diagnostics) != 3 {
-		t.Fatalf("diagnostics = %#v, want three independent action failures", report.Diagnostics)
-	}
-	wantResults := map[int]string{1: compatibility.Passed, 2: compatibility.Failed}
-	missingFailures := 0
-	for _, result := range report.Actions {
-		switch result.Job {
-		case "gha-duplicate":
-			if result.Result != wantResults[result.Step] {
-				t.Fatalf("duplicate action result = %#v", result)
-			}
-		case "gha-missing":
-			if result.Result == compatibility.Failed {
-				missingFailures++
-			}
-		}
-	}
-	if missingFailures != 2 {
-		t.Fatalf("actions = %#v, want both missing invocations failed", report.Actions)
-	}
-	for _, diagnostic := range report.Diagnostics {
-		if diagnostic.Instance == "" || diagnostic.Step == 0 || diagnostic.Code != compiler.CodeActionResolution || !strings.Contains(diagnostic.Message, `Action "`) || !strings.Contains(diagnostic.Message, "unsupported") {
-			t.Fatalf("diagnostic lacks invocation identity: %#v", diagnostic)
-		}
-	}
-}
-
-func TestValidateReportsIndependentWorkflowAndEventSyntaxFailures(t *testing.T) {
-	root := t.TempDir()
-	workflowPath := filepath.Join(root, "invalid.yml")
-	eventPath := filepath.Join(root, "event.json")
-	if err := os.WriteFile(workflowPath, []byte("jobs: [\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(eventPath, []byte("{\"provider\":"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"validate", "--format", "json", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 1 {
-		t.Fatalf("Run() code = %d, want 1; stderr = %q", code, stderr.String())
-	}
-	var report compatibility.ProcessingReport
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	results := map[string]string{}
-	for _, stage := range report.Stages {
-		results[stage.ID] = stage.Result
-	}
-	if results[stageWorkflowParsing] != compatibility.Failed || results[stageEventValidation] != compatibility.Failed || results[stageGraph] != compatibility.NotEvaluated {
-		t.Fatalf("stage results = %#v", results)
-	}
-	if len(report.Diagnostics) != 2 || report.Diagnostics[0].Category != "syntax" || report.Diagnostics[1].Category != "environment" {
-		t.Fatalf("diagnostics = %#v", report.Diagnostics)
-	}
-
-	stdout.Reset()
-	stderr.Reset()
-	validEventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	if code := Run([]string{"validate", "--format", "json", "--event-path", validEventPath, workflowPath}, &stdout, &stderr, "dev"); code != 1 {
-		t.Fatalf("Run() with valid event code = %d, want 1; stderr = %q", code, stderr.String())
-	}
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	results = map[string]string{}
-	for _, stage := range report.Stages {
-		results[stage.ID] = stage.Result
-	}
-	if results[stageWorkflowParsing] != compatibility.Failed || results[stageEventValidation] != compatibility.Passed || results[stageGraph] != compatibility.NotEvaluated {
-		t.Fatalf("stage results with valid event = %#v", results)
-	}
-	if len(report.Diagnostics) != 1 || report.Diagnostics[0].Category != "syntax" {
-		t.Fatalf("diagnostics with valid event = %#v", report.Diagnostics)
 	}
 }
 
@@ -1439,90 +1186,6 @@ func TestProcessingReportRedactsEventDerivedMatrixKeys(t *testing.T) {
 	}
 }
 
-func TestCompileAggregatesIndependentPlanFailuresWithoutPipeline(t *testing.T) {
-	workflowPath := filepath.Join(t.TempDir(), "plans.yml")
-	workflow := []byte(`on: push
-permissions: {}
-jobs:
-  first:
-    runs-on: ubuntu-latest
-    steps:
-      - run: echo '${{ secrets.GITHUB_TOKEN }}'
-  second:
-    runs-on: ubuntu-latest
-    steps:
-      - run: echo '${{ secrets.GITHUB_TOKEN }}'
-  dependent:
-    needs: first
-    runs-on: ubuntu-latest
-    steps:
-      - run: true
-`)
-	if err := os.WriteFile(workflowPath, workflow, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"compile", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 1 {
-		t.Fatalf("Run() code = %d, want 1; stderr = %q", code, stderr.String())
-	}
-	if stdout.Len() != 0 {
-		t.Fatalf("compile emitted partial pipeline: %q", stdout.String())
-	}
-	if got := strings.Count(stderr.String(), "[E_PLAN_CONSTRUCTION]"); got != 2 {
-		t.Fatalf("plan diagnostics = %d, want 2; report = %q", got, stderr.String())
-	}
-	for _, want := range []string{"job first/gha-first: failed", "job second/gha-second: failed", "job dependent/gha-dependent: not-evaluated", "Pipeline generation: not-evaluated"} {
-		if !strings.Contains(stderr.String(), want) {
-			t.Fatalf("report = %q, want %q", stderr.String(), want)
-		}
-	}
-}
-
-func TestHostedReportAdmitsIndependentContainerJobs(t *testing.T) {
-	workflowPath := filepath.Join(t.TempDir(), "containers.yml")
-	workflow := []byte(`on: push
-jobs:
-  first:
-    runs-on: ubuntu-latest
-    container: alpine:3.20
-    steps:
-      - run: true
-  second:
-    runs-on: ubuntu-latest
-    container: alpine:3.20
-    steps:
-      - run: true
-`)
-	if err := os.WriteFile(workflowPath, workflow, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	args := []string{"validate", "--profile", "hosted-tokenless", "--format", "json", "--event-path", eventPath, workflowPath}
-	if code := Run(args, &stdout, &stderr, "dev"); code != 0 {
-		t.Fatalf("Run() code = %d, want 0; stderr = %q", code, stderr.String())
-	}
-	var report compatibility.ProcessingReport
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	instances := map[string]string{}
-	for _, diagnostic := range report.Diagnostics {
-		if diagnostic.Code == "E_PROFILE" {
-			t.Fatalf("unexpected profile diagnostic: %#v", diagnostic)
-		}
-	}
-	for _, job := range report.Jobs {
-		if job.Instance != "" {
-			instances[job.Instance] = job.Result
-		}
-	}
-	if report.Result != "admitted" || report.Admission.Result != "admitted" || instances["gha-first"] != compatibility.Passed || instances["gha-second"] != compatibility.Passed {
-		t.Fatalf("report = %#v", report)
-	}
-}
-
 func TestProcessingReportRetainsReusableCalleeJobsAfterFailure(t *testing.T) {
 	root := t.TempDir()
 	workflowRoot := filepath.Join(root, ".github", "workflows")
@@ -1611,34 +1274,6 @@ jobs:
 	}
 	if len(report.Diagnostics) != 1 || report.Diagnostics[0].Job != "delegated" || report.Diagnostics[0].Stage != stageGraph {
 		t.Fatalf("diagnostics = %#v", report.Diagnostics)
-	}
-}
-
-func TestHostedProfileReportsNarrowedReusableWorkflowTokenWarning(t *testing.T) {
-	root := t.TempDir()
-	workflowRoot := filepath.Join(root, ".github", "workflows")
-	if err := os.MkdirAll(workflowRoot, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	workflowPath := filepath.Join(workflowRoot, "caller.yml")
-	if err := os.WriteFile(workflowPath, []byte("on: push\npermissions:\n  contents: write\njobs:\n  delegated:\n    uses: ./.github/workflows/reusable.yml\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(workflowRoot, "reusable.yml"), []byte("on: workflow_call\npermissions:\n  contents: read\njobs:\n  token:\n    runs-on: ubuntu-latest\n    steps:\n      - run: echo '${{ secrets.GITHUB_TOKEN }}'\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"validate", "--profile", "hosted", "--format", "json", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 0 {
-		t.Fatalf("Run() code = %d, want 0; stderr = %q", code, stderr.String())
-	}
-	var report compatibility.ProcessingReport
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	if len(report.Diagnostics) != 1 || report.Diagnostics[0].Code != "W_REUSABLE_WORKFLOW_TOKEN_USES_ROOT_PERMISSIONS" || report.Diagnostics[0].Level != "warning" {
-		t.Fatalf("diagnostics = %#v, want reusable workflow token warning", report.Diagnostics)
 	}
 }
 
@@ -1799,72 +1434,6 @@ jobs:
 		pipeline := string(runner.commands[len(runner.commands)-1].stdin)
 		if strings.Count(pipeline, "concurrency_group:") != 2 || strings.Contains(pipeline, "agents:") {
 			t.Fatalf("uploaded commands = %#v", runner.commands)
-		}
-	})
-}
-
-func TestValidatePublishesProcessingDiagnosticsInBuildkite(t *testing.T) {
-	t.Setenv("BUILDKITE", "true")
-	t.Setenv("BUILDKITE_JOB_ID", cliTestJobID)
-
-	t.Run("error", func(t *testing.T) {
-		workflowPath := filepath.Join(t.TempDir(), "unsupported.yml")
-		if err := os.WriteFile(workflowPath, []byte("on: push\njobs:\n  test:\n    runs-on: windows-latest\n    steps: [{run: true}]\n"), 0o600); err != nil {
-			t.Fatal(err)
-		}
-		runner := &cliCaptureRunner{}
-		var stdout, stderr bytes.Buffer
-		if code := run([]string{"validate", "--format", "json", workflowPath}, &stdout, &stderr, "dev", runner); code != 1 {
-			t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-		}
-		if len(runner.commands) != 1 {
-			t.Fatalf("commands = %#v, want one annotation", runner.commands)
-		}
-		annotation := runner.commands[0]
-		if len(annotation.args) != 9 || annotation.args[0] != "annotate" || annotation.args[2] != "job" || annotation.args[4] != cliTestJobID || !strings.HasPrefix(annotation.args[6], processingAnnotationContext+"-") || annotation.args[8] != "error" {
-			t.Fatalf("annotation args = %#v", annotation.args)
-		}
-		for _, want := range []string{
-			`<h2 class="h4 mb2">Workflow could not be run</h2>`,
-			`<p><strong>Runner label &#34;windows-latest&#34; requires Windows, which is unsupported.`,
-			`<summary>Diagnostic detail</summary>`,
-			`Supported runner labels: ubuntu-22.04, ubuntu-24.04, ubuntu-latest.`,
-			"Job <code>test</code>",
-		} {
-			if !strings.Contains(string(annotation.stdin), want) {
-				t.Fatalf("annotation = %q, want %q", annotation.stdin, want)
-			}
-		}
-		for _, unwanted := range []string{"E_EXPRESSION_INVALID", "stage:", "instance:", "gha-test"} {
-			if strings.Contains(string(annotation.stdin), unwanted) {
-				t.Fatalf("annotation = %q, does not want %q", annotation.stdin, unwanted)
-			}
-		}
-		var report compatibility.ProcessingReport
-		if err := json.Unmarshal(stdout.Bytes(), &report); err != nil || report.Result != "incompatible" {
-			t.Fatalf("report = %#v, error = %v", report, err)
-		}
-	})
-
-	t.Run("warning publication failure is non-fatal", func(t *testing.T) {
-		workflowPath := filepath.Join(t.TempDir(), "warning.yml")
-		workflow := []byte("on: push\nconcurrency:\n  group: ci\n  cancel-in-progress: true\njobs:\n  test:\n    runs-on: ubuntu-latest\n    steps: [{run: true}]\n")
-		if err := os.WriteFile(workflowPath, workflow, 0o600); err != nil {
-			t.Fatal(err)
-		}
-		runner := &cliCaptureRunner{failAnnotation: true}
-		var stdout, stderr bytes.Buffer
-		if code := run([]string{"validate", workflowPath}, &stdout, &stderr, "dev", runner); code != 0 {
-			t.Fatalf("run() code = %d, stderr = %q", code, stderr.String())
-		}
-		if len(runner.commands) != 1 || runner.commands[0].args[8] != "warning" {
-			t.Fatalf("commands = %#v, want one warning annotation", runner.commands)
-		}
-		if body := string(runner.commands[0].stdin); !strings.Contains(body, `<h2 class="h4 mb2">GitHub Actions workflow diagnostics</h2>`) || strings.Contains(body, "Workflow could not be run") {
-			t.Fatalf("warning annotation = %q", body)
-		}
-		if !strings.Contains(stderr.String(), "warning: processing annotation: annotation unavailable") {
-			t.Fatalf("stderr = %q", stderr.String())
 		}
 	})
 }
@@ -2529,98 +2098,6 @@ runs:
 	}
 	if len(runner.commands) == 0 {
 		t.Fatal("conditional action default did not upload a pipeline")
-	}
-}
-
-func TestValidateHostedProfileResolvesActionsWithoutClaimingRuntime(t *testing.T) {
-	root := t.TempDir()
-	workflowPath := filepath.Join(root, ".github", "workflows", "action.yml")
-	actionPath := filepath.Join(root, ".github", "actions", "local")
-	if err := os.MkdirAll(filepath.Dir(workflowPath), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.MkdirAll(actionPath, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(workflowPath, []byte("on: push\njobs:\n  action:\n    runs-on: ubuntu-latest\n    steps:\n      - uses: ./.github/actions/local\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(actionPath, "action.yml"), []byte("runs:\n  using: node24\n  main: main.js\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(actionPath, "main.js"), []byte("console.log('local action')\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	t.Setenv("BUILDKITE_GHA_NODE20", writeFakeNode(t, root, 20))
-	t.Setenv("BUILDKITE_GHA_NODE24", writeFakeNode(t, root, 24))
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"validate", "--profile", "hosted-tokenless", "--format", "json", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 0 {
-		t.Fatalf("Run() code = %d, stderr = %q", code, stderr.String())
-	}
-	var report compatibility.ProcessingReport
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	if report.Result != "admitted" || len(report.Diagnostics) != 1 || report.Diagnostics[0].Code != "W_ACTION_RUNTIME_UNKNOWN" {
-		t.Fatalf("profile report = %#v", report)
-	}
-
-	t.Setenv("BUILDKITE_GHA_NODE20", filepath.Join(root, "missing-node20"))
-	t.Setenv("BUILDKITE_GHA_TARGET_QUEUE", "not a queue")
-	stdout.Reset()
-	stderr.Reset()
-	if code := Run([]string{"validate", "--profile", "hosted-tokenless", "--format", "json", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 0 {
-		t.Fatalf("environment override affected production profile: code = %d; stderr = %q", code, stderr.String())
-	}
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	if report.Result != "admitted" {
-		t.Fatalf("environment profile report = %#v", report)
-	}
-}
-
-func TestValidateHostedProfileAdmitsOrdinarySecretsAfterCompile(t *testing.T) {
-	workflowPath := filepath.Join(t.TempDir(), "secret.yml")
-	if err := os.WriteFile(workflowPath, []byte("on: push\njobs:\n  secret:\n    runs-on: ubuntu-latest\n    env:\n      TOKEN: ${{ secrets.TOKEN }}\n    steps:\n      - run: true\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"validate", "--profile", "hosted-tokenless", "--format", "json", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 0 {
-		t.Fatalf("Run() code = %d, want 0; stderr = %q", code, stderr.String())
-	}
-	var report compatibility.ProcessingReport
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	if report.Result != "admitted" || report.Compile.Result != "compilable" || report.Admission.Result != "admitted" || len(report.Diagnostics) != 0 {
-		t.Fatalf("profile report = %#v", report)
-	}
-}
-
-func TestValidateHostedProfileAdmitsImplicitReadOnlyWorkflowToken(t *testing.T) {
-	root := t.TempDir()
-	workflowPath := filepath.Join(root, ".github", "workflows", "token.yml")
-	if err := os.MkdirAll(filepath.Dir(workflowPath), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	workflow := []byte("on: push\njobs:\n  token:\n    runs-on: ubuntu-latest\n    steps:\n      - run: echo '${{ secrets.GITHUB_TOKEN }}'\n")
-	if err := os.WriteFile(workflowPath, workflow, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	eventPath := filepath.Join("..", "..", "testdata", "smoke", "events", "push.json")
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"validate", "--profile", "hosted", "--format", "json", "--event-path", eventPath, workflowPath}, &stdout, &stderr, "dev"); code != 0 {
-		t.Fatalf("Run() code = %d, stderr = %q", code, stderr.String())
-	}
-	var report compatibility.ProcessingReport
-	if err := json.Unmarshal(stdout.Bytes(), &report); err != nil {
-		t.Fatal(err)
-	}
-	if report.Result != "admitted" || report.Compile.Result != "compilable" || report.Admission.Result != "admitted" || len(report.Diagnostics) != 0 {
-		t.Fatalf("profile report = %#v", report)
 	}
 }
 
@@ -4539,271 +4016,6 @@ func TestRunUploadRejectsAllReusableExplicitPaths(t *testing.T) {
 	}
 }
 
-func TestActionSourceAuthenticationUsesDedicatedTokenAndFallsBackAnonymously(t *testing.T) {
-	const token = "ghs_action_source"
-	provider := &cliActionSourceTokenProvider{token: token}
-	redactor := &cliRedactor{}
-	authentication := &actionSourceAuthentication{provider: provider, redactor: redactor}
-	option := authentication.option("buildkite/buildkite-gha")
-	if option == nil || provider.calls != 0 || len(redactor.values) != 0 {
-		t.Fatalf("authentication = option %v, provider %#v, redactions %#v", option != nil, provider, redactor.values)
-	}
-
-	var requests int
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		requests++
-		if r.Header.Get("Authorization") != "Bearer "+token {
-			t.Errorf("Authorization = %q", r.Header.Get("Authorization"))
-		}
-		if r.URL.Path == "/repos/o/r" {
-			_, _ = io.WriteString(w, `{"private":false,"visibility":"public"}`)
-			return
-		}
-		_, _ = io.WriteString(w, `{"object":{"type":"commit","sha":"0123456789abcdef0123456789abcdef01234567"}}`)
-	}))
-	defer server.Close()
-	resolver, err := actionsource.NewResolver(server.Client(), option, actionsource.WithTestEndpoints(server.URL))
-	if err != nil {
-		t.Fatal(err)
-	}
-	exact, err := actionsource.Parse("o/r@0123456789abcdef0123456789abcdef01234567")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, err := resolver.Resolve(context.Background(), exact); err != nil || requests != 0 || provider.calls != 0 || len(redactor.values) != 0 {
-		t.Fatalf("exact Resolve() error/requests/provider/redactions = %v / %d / %d / %#v", err, requests, provider.calls, redactor.values)
-	}
-	ref, err := actionsource.Parse("o/r@v1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, err := resolver.Resolve(context.Background(), ref); err != nil || requests != 2 || provider.calls != 1 || provider.repository != "buildkite/buildkite-gha" || !slices.Equal(redactor.values, []string{token}) {
-		t.Fatalf("Resolve() error/requests = %v / %d", err, requests)
-	}
-	second, _ := actionsource.Parse("o/r@v2")
-	if _, err := resolver.Resolve(context.Background(), second); err != nil || provider.calls != 1 || !slices.Equal(redactor.values, []string{token}) {
-		t.Fatalf("second Resolve() error/provider/redactions = %v / %d / %#v", err, provider.calls, redactor.values)
-	}
-
-	for _, test := range []struct {
-		name          string
-		provider      *cliActionSourceTokenProvider
-		redactor      *cliRedactor
-		cancelContext bool
-		wantContext   bool
-	}{
-		{name: "unavailable", redactor: &cliRedactor{}},
-		{name: "mint failure", provider: &cliActionSourceTokenProvider{err: errors.New("secret backend details")}, redactor: &cliRedactor{}},
-		{name: "redaction failure", provider: &cliActionSourceTokenProvider{token: token}, redactor: &cliRedactor{err: errors.New("secret backend details")}},
-		{name: "mint client timeout", provider: &cliActionSourceTokenProvider{err: context.DeadlineExceeded}, redactor: &cliRedactor{}},
-		{name: "redaction client timeout", provider: &cliActionSourceTokenProvider{token: token}, redactor: &cliRedactor{err: context.DeadlineExceeded}},
-		{name: "pre-cancelled while unavailable", redactor: &cliRedactor{}, cancelContext: true, wantContext: true},
-		{name: "mint cancellation", provider: &cliActionSourceTokenProvider{err: context.Canceled}, redactor: &cliRedactor{}, wantContext: true},
-		{name: "redaction cancellation", provider: &cliActionSourceTokenProvider{token: token}, redactor: &cliRedactor{err: context.Canceled}, wantContext: true},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			var warnings bytes.Buffer
-			ctx := context.Background()
-			if test.cancelContext {
-				var cancel context.CancelFunc
-				ctx, cancel = context.WithCancel(ctx)
-				cancel()
-			}
-			authentication := &actionSourceAuthentication{redactor: test.redactor, warnings: &warnings}
-			if test.provider != nil {
-				authentication.provider = test.provider
-			}
-			gotToken, err := authentication.token(ctx, "buildkite/buildkite-gha")
-			if test.wantContext {
-				if !errors.Is(err, context.Canceled) || gotToken != "" || warnings.Len() != 0 {
-					t.Fatalf("token/error/warnings = %q / %v / %q", gotToken, err, warnings.String())
-				}
-			} else if err != nil || gotToken != "" {
-				t.Fatalf("token/error = %q / %v, want anonymous fallback", gotToken, err)
-			} else if !strings.Contains(warnings.String(), "resolving mutable public action references anonymously") || strings.Contains(warnings.String(), token) || strings.Contains(warnings.String(), "backend details") {
-				t.Fatalf("fallback warning = %q, want sanitized observable warning", warnings.String())
-			}
-		})
-	}
-}
-
-func TestActionSourceAuthenticationReusesOneTokenAcrossConcurrentWorkflowResolvers(t *testing.T) {
-	const token = "ghs_action_source"
-	provider := &cliActionSourceTokenProvider{token: token}
-	redactor := &cliRedactor{}
-	authentication := &actionSourceAuthentication{provider: provider, redactor: redactor}
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path == "/repos/o/r" {
-			_, _ = io.WriteString(w, `{"private":false,"visibility":"public"}`)
-			return
-		}
-		_, _ = io.WriteString(w, `{"object":{"type":"commit","sha":"0123456789abcdef0123456789abcdef01234567"}}`)
-	}))
-	defer server.Close()
-
-	const resolverCount = 11
-	resolvers := make([]*actionsource.Resolver, resolverCount)
-	for i := range resolvers {
-		resolver, err := actionsource.NewResolver(server.Client(), authentication.option("buildkite/buildkite-gha"), actionsource.WithTestEndpoints(server.URL))
-		if err != nil {
-			t.Fatal(err)
-		}
-		resolvers[i] = resolver
-	}
-	ref, err := actionsource.Parse("o/r@v1")
-	if err != nil {
-		t.Fatal(err)
-	}
-	var wg sync.WaitGroup
-	errs := make(chan error, resolverCount)
-	for _, resolver := range resolvers {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			_, err := resolver.Resolve(context.Background(), ref)
-			errs <- err
-		}()
-	}
-	wg.Wait()
-	close(errs)
-	for err := range errs {
-		if err != nil {
-			t.Fatal(err)
-		}
-	}
-	if provider.calls != 1 || !slices.Equal(redactor.values, []string{token}) {
-		t.Fatalf("provider calls/redactions = %d / %#v, want one importer-job mint and redaction", provider.calls, redactor.values)
-	}
-}
-
-func TestHostedLocalActionDoesNotProvisionSourceToken(t *testing.T) {
-	root := t.TempDir()
-	workflowPath := filepath.Join(root, ".github", "workflows", "local.yml")
-	actionPath := filepath.Join(root, ".github", "actions", "local")
-	if err := os.MkdirAll(filepath.Dir(workflowPath), 0o755); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.MkdirAll(actionPath, 0o755); err != nil {
-		t.Fatal(err)
-	}
-	workflowSource := []byte("on: push\njobs:\n  local:\n    runs-on: ubuntu-latest\n    steps:\n      - uses: ./.github/actions/local\n")
-	if err := os.WriteFile(workflowPath, workflowSource, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(filepath.Join(actionPath, "action.yml"), []byte("runs:\n  using: composite\n  steps:\n    - shell: sh\n      run: true\n"), 0o600); err != nil {
-		t.Fatal(err)
-	}
-	eventSource, err := os.ReadFile(filepath.Join("..", "..", "testdata", "smoke", "events", "push.json"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	originEventSource := bytes.Replace(eventSource, []byte(`"provider": "github"`), []byte(`"provider": "cursor-origin"`), 1)
-	originEventSource = bytes.Replace(originEventSource, []byte(`"owner": "buildkite"`), []byte(`"owner": "acme_team"`), 1)
-	originEventSource = bytes.Replace(originEventSource, []byte(`"name": "buildkite-gha"`), []byte(`"name": "widgets"`), 1)
-	originEventSource = bytes.Replace(originEventSource, []byte(`"clone_url": "https://github.com/buildkite/buildkite-gha.git"`), []byte(`"clone_url": "https://origin.cursor.com/git/acme_team/widgets.git"`), 1)
-	for _, test := range []struct {
-		name  string
-		event []byte
-	}{
-		{name: "GitHub", event: eventSource},
-		{name: "Origin repository with GitHub-incompatible namespace", event: originEventSource},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			provider := &cliActionSourceTokenProvider{token: "must-not-be-minted"}
-			redactor := &cliRedactor{}
-			var warnings bytes.Buffer
-			authentication := &actionSourceAuthentication{provider: provider, redactor: redactor, warnings: &warnings}
-			if _, err := compileHosted(context.Background(), workflowPath, workflowSource, test.event, "dev", "sha256:"+strings.Repeat("0", 64), "importer", "", nil, nil, authentication); err != nil {
-				t.Fatal(err)
-			}
-			if provider.calls != 0 || len(redactor.values) != 0 || warnings.Len() != 0 {
-				t.Fatalf("local action provisioned source credential: calls %d, redactions %#v, warnings %q", provider.calls, redactor.values, warnings.String())
-			}
-		})
-	}
-}
-
-func TestCompileHostedRequiresExplicitMacOSQueueAndRuntime(t *testing.T) {
-	workflow := []byte("on: push\njobs:\n  macos:\n    runs-on: macos-15\n    steps:\n      - run: echo macos\n")
-	event, err := os.ReadFile(filepath.Join("..", "..", "testdata", "smoke", "events", "push.json"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	compilerDigest := "sha256:" + strings.Repeat("0", 64)
-	darwinDigest := "sha256:" + strings.Repeat("1", 64)
-	_, err = compileHosted(context.Background(), "macos.yml", workflow, event, "0.0.0-test", compilerDigest, "importer", "", nil, nil, nil)
-	if err == nil || !strings.Contains(err.Error(), `runner label "macos-15" is not mapped by policy`) {
-		t.Fatalf("missing macOS queue error = %v", err)
-	}
-	macOSTarget := map[string]compiler.RunnerTarget{"macos-15": {Queue: "macos", Platform: compiler.PlatformDarwinARM64}}
-	_, err = compileHosted(context.Background(), "macos.yml", workflow, event, "0.0.0-test", compilerDigest, "importer", "", macOSTarget, nil, nil)
-	if err == nil || !strings.Contains(err.Error(), "no runtime distribution configured for darwin/arm64") {
-		t.Fatalf("missing macOS runtime error = %v", err)
-	}
-	compiled, err := compileHosted(context.Background(), "macos.yml", workflow, event, "0.0.0-test", compilerDigest, "importer", "", macOSTarget, map[compiler.Platform]string{compiler.PlatformDarwinARM64: darwinDigest}, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(compiled.Bundle.Plans) != 1 || compiled.Bundle.Plans[0].Job.Target.Queue != "macos" || compiled.Bundle.Plans[0].Job.RuntimeDistributionDigest() != darwinDigest || !bytes.Contains(compiled.Bundle.Pipeline, []byte("queue: \"macos\"")) {
-		t.Fatalf("macOS compilation = %#v\n%s", compiled.Bundle.Plans, compiled.Bundle.Pipeline)
-	}
-}
-
-func TestCompileHostedDefaultsMacOSLatestAliasToHostedQueue(t *testing.T) {
-	workflow := []byte("on: push\njobs:\n  macos:\n    runs-on: macOS-latest\n    steps:\n      - run: echo macos\n")
-	event, err := os.ReadFile(filepath.Join("..", "..", "testdata", "smoke", "events", "push.json"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	compilerDigest := "sha256:" + strings.Repeat("0", 64)
-	darwinDigest := "sha256:" + strings.Repeat("1", 64)
-	_, err = compileHosted(context.Background(), "macos.yml", workflow, event, "0.0.0-test", compilerDigest, "importer", "", nil, nil, nil)
-	if err == nil || !strings.Contains(err.Error(), "no runtime distribution configured for darwin/arm64") {
-		t.Fatalf("unmapped macos-latest without Darwin runtime error = %v", err)
-	}
-	darwinRuntime := map[compiler.Platform]string{compiler.PlatformDarwinARM64: darwinDigest}
-	compiled, err := compileHosted(context.Background(), "macos.yml", workflow, event, "0.0.0-test", compilerDigest, "importer", "", nil, darwinRuntime, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(compiled.Bundle.Plans) != 1 || compiled.Bundle.Plans[0].Job.Target.Queue != defaultMacOSRunnerQueue || compiled.Bundle.Plans[0].Job.RuntimeDistributionDigest() != darwinDigest || !bytes.Contains(compiled.Bundle.Pipeline, []byte("queue: \"macos-medium\"")) {
-		t.Fatalf("default macos-latest compilation = %#v\n%s", compiled.Bundle.Plans, compiled.Bundle.Pipeline)
-	}
-	override := map[string]compiler.RunnerTarget{"macos-latest": {Queue: "custom-macos", Platform: compiler.PlatformDarwinARM64}}
-	compiled, err = compileHosted(context.Background(), "macos.yml", workflow, event, "0.0.0-test", compilerDigest, "importer", "", override, darwinRuntime, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(compiled.Bundle.Plans) != 1 || compiled.Bundle.Plans[0].Job.Target.Queue != "custom-macos" || !bytes.Contains(compiled.Bundle.Pipeline, []byte("queue: \"custom-macos\"")) {
-		t.Fatalf("overridden macos-latest compilation = %#v\n%s", compiled.Bundle.Plans, compiled.Bundle.Pipeline)
-	}
-}
-
-func TestJobScopedActionSourceAuthenticationIgnoresAmbientGitHubTokens(t *testing.T) {
-	t.Setenv("GH_TOKEN", "ignored")
-	t.Setenv("GITHUB_TOKEN", "ignored")
-	t.Setenv("BUILDKITE_GHA_GITHUB_TOKEN", "ignored")
-	t.Setenv("BUILDKITE_AGENT_ENDPOINT", "")
-	t.Setenv("BUILDKITE_JOB_ID", "")
-	t.Setenv("BUILDKITE_AGENT_ACCESS_TOKEN", "")
-	var warnings bytes.Buffer
-	authentication := importerJobActionSourceAuthentication(&warnings)
-	token, err := authentication.token(context.Background(), "buildkite/buildkite-gha")
-	if err != nil || token != "" || authentication.provider != nil || !strings.Contains(warnings.String(), "authentication is unavailable") {
-		t.Fatalf("ambient GitHub token authentication = provider %v, token %q, error %v, warnings %q", authentication.provider != nil, token, err, warnings.String())
-	}
-
-	server := httptest.NewServer(http.NotFoundHandler())
-	defer server.Close()
-	t.Setenv("BUILDKITE_AGENT_ENDPOINT", server.URL)
-	t.Setenv("BUILDKITE_JOB_ID", cliTestJobID)
-	t.Setenv("BUILDKITE_AGENT_ACCESS_TOKEN", "job-token")
-	if authentication := importerJobActionSourceAuthentication(io.Discard); authentication.provider == nil {
-		t.Fatal("job-scoped Agent configuration did not configure action-source authentication")
-	}
-}
-
 func TestRunUploadUsesExplicitTargetQueueAndRunnerUserDefault(t *testing.T) {
 	requireImporterHost(t)
 	workflowPath := filepath.Join("..", "..", "testdata", "smoke", ".github", "workflows", "shell.yml")
@@ -4950,62 +4162,6 @@ func TestRunUploadUsesExplicitRuntimeImage(t *testing.T) {
 		if step.Image != image || !strings.Contains(step.Command, "--hosted-tool-cache") {
 			t.Fatalf("runtime image step = %#v", step)
 		}
-	}
-}
-
-func TestRunnerProfileDoesNotAffectOtherLinuxLabels(t *testing.T) {
-	workflow := []byte(`on: push
-jobs:
-  selected:
-    runs-on: ubuntu-latest
-    steps:
-      - run: echo selected
-  default:
-    runs-on: ubuntu-22.04
-    steps:
-      - run: echo default
-`)
-	event, err := os.ReadFile(filepath.Join("..", "..", "testdata", "smoke", "events", "push.json"))
-	if err != nil {
-		t.Fatal(err)
-	}
-	image := "buildkite.namespace-images.com/agent-base@sha256:" + strings.Repeat("0", 64)
-	targets := map[string]compiler.RunnerTarget{
-		"ubuntu-latest": {Queue: "hosted", Platform: compiler.PlatformLinuxAMD64, Image: image},
-	}
-	compiled, err := compileHosted(context.Background(), "profiles.yml", workflow, event, "dev", "sha256:"+strings.Repeat("1", 64), "importer", "", targets, nil, nil)
-	if err != nil {
-		t.Fatal(err)
-	}
-	var pipeline struct {
-		Steps []struct {
-			Key     string            `yaml:"key"`
-			Image   string            `yaml:"image"`
-			Agents  map[string]string `yaml:"agents"`
-			Command string            `yaml:"command"`
-		} `yaml:"steps"`
-	}
-	if err := yaml.Unmarshal(compiled.Bundle.Pipeline, &pipeline); err != nil {
-		t.Fatal(err)
-	}
-	steps := make(map[string]struct {
-		Image   string
-		Agents  map[string]string
-		Command string
-	}, len(pipeline.Steps))
-	for _, step := range pipeline.Steps {
-		steps[step.Key] = struct {
-			Image   string
-			Agents  map[string]string
-			Command string
-		}{Image: step.Image, Agents: step.Agents, Command: step.Command}
-	}
-	selected, defaulted := steps["gha-selected"], steps["gha-default"]
-	if selected.Image != image || selected.Agents["queue"] != "hosted" || !strings.Contains(selected.Command, "--hosted-tool-cache") {
-		t.Fatalf("selected profile step = %#v", selected)
-	}
-	if defaulted.Image != defaultJammyRunnerImage || len(defaulted.Agents) != 0 || !strings.Contains(defaulted.Command, "--hosted-tool-cache") {
-		t.Fatalf("unmapped Linux step = %#v, want default targeting with the Jammy hosted-toolchains image", defaulted)
 	}
 }
 
@@ -5417,326 +4573,6 @@ func TestRunUploadFailsClosedBeforePipeline(t *testing.T) {
 	}
 }
 
-func TestUnprivilegedUploadRejectsCapabilities(t *testing.T) {
-	tests := []struct {
-		capability string
-		want       string
-	}{
-		{capability: "privileged-container", want: `Job "protected" requires unsupported hosted runtime capability "privileged-container". Remove the requirement or use a runtime profile that supports it.`},
-		{capability: "future-capability", want: `Job "protected" requires unsupported hosted runtime capability "future-capability". Remove the requirement or use a runtime profile that supports it.`},
-	}
-	for _, test := range tests {
-		capability := test.capability
-		bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-			Workflow:             plan.Workflow{LogicalJobID: "protected"},
-			RequiredCapabilities: []string{capability},
-		}}}}
-		err := validateUnprivilegedBundle(bundle)
-		var finding *compiler.ProcessingFinding
-		if err == nil || !errors.As(err, &finding) || finding.Message != test.want || finding.Detail != "" {
-			t.Fatalf("validateUnprivilegedBundle(%q) error = %v, want capability rejection", capability, err)
-		}
-	}
-}
-
-func TestUnprivilegedUploadAdmitsSecretsCapability(t *testing.T) {
-	bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-		Workflow:             plan.Workflow{LogicalJobID: "release"},
-		RequiredCapabilities: []string{"secrets"},
-		RequiredSecrets:      []string{"HOMEBREW_TAP_GITHUB_TOKEN"},
-	}}}}
-	if err := validateUnprivilegedBundle(bundle); err != nil {
-		t.Fatalf("validateUnprivilegedBundle() error = %v", err)
-	}
-}
-
-func TestUnprivilegedUploadAdmitsOnlyCompilerVerifiedCheckoutCredentials(t *testing.T) {
-	job := plan.Job{
-		Workflow:             plan.Workflow{LogicalJobID: "checkout"},
-		RequiredCapabilities: []string{"network", "provider-token-read"},
-	}
-	for _, test := range []struct {
-		name          string
-		authorization compiler.PlanAuthorization
-		wantError     bool
-	}{
-		{name: "verified adapter", authorization: compiler.PlanAuthorization{ProviderTokenReadCapabilitySources: []string{"checkout-adapter"}}},
-		{name: "missing provenance", wantError: true},
-		{name: "broadened provenance", authorization: compiler.PlanAuthorization{ProviderTokenReadCapabilitySources: []string{"checkout-adapter", "javascript-action"}}, wantError: true},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: job, Authorization: test.authorization}}}
-			err := validateUnprivilegedBundle(bundle)
-			if test.wantError && (err == nil || !strings.Contains(err.Error(), "unsupported GitHub checkout credentials")) {
-				t.Fatalf("validateUnprivilegedBundle() error = %v", err)
-			}
-			if !test.wantError && err != nil {
-				t.Fatalf("validateUnprivilegedBundle() error = %v", err)
-			}
-		})
-	}
-}
-
-func TestUnprivilegedUploadAdmitsOnlyCompilerVerifiedWorkflowToken(t *testing.T) {
-	job := plan.Job{
-		Workflow:             plan.Workflow{Path: ".github/workflows/comment.yml", LogicalJobID: "comment"},
-		RequiredCapabilities: []string{"provider-token-write"},
-		GitHubToken:          &plan.GitHubToken{Workflow: "comment.yml", Permissions: map[string]string{"pull_requests": "write"}},
-	}
-	for _, test := range []struct {
-		name          string
-		authorization compiler.PlanAuthorization
-		wantError     bool
-	}{
-		{name: "verified policy", authorization: compiler.PlanAuthorization{ProviderTokenWriteCapabilitySources: []string{"effective-permissions"}, WorkflowTokenPolicyFilename: "comment.yml"}},
-		{name: "missing provenance", wantError: true},
-		{name: "missing policy", authorization: compiler.PlanAuthorization{ProviderTokenWriteCapabilitySources: []string{"effective-permissions"}}, wantError: true},
-		{name: "mismatched policy", authorization: compiler.PlanAuthorization{ProviderTokenWriteCapabilitySources: []string{"effective-permissions"}, WorkflowTokenPolicyFilename: "other.yml"}, wantError: true},
-		{name: "broadened provenance", authorization: compiler.PlanAuthorization{ProviderTokenWriteCapabilitySources: []string{"effective-permissions", "step-input"}, WorkflowTokenPolicyFilename: "comment.yml"}, wantError: true},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: job, Authorization: test.authorization}}}
-			err := validateUnprivilegedBundle(bundle)
-			if test.wantError && (err == nil || !strings.Contains(err.Error(), "GITHUB_TOKEN")) {
-				t.Fatalf("validateUnprivilegedBundle() error = %v", err)
-			}
-			if !test.wantError && err != nil {
-				t.Fatalf("validateUnprivilegedBundle() error = %v", err)
-			}
-		})
-	}
-}
-
-func TestGitHubTokenAdmissionDiagnosticSeparatesGuidanceFromDetail(t *testing.T) {
-	artifact := compiler.PlanArtifact{
-		Job: plan.Job{
-			Workflow:    plan.Workflow{LogicalJobID: "build"},
-			GitHubToken: &plan.GitHubToken{Workflow: "build.yml", Permissions: map[string]string{"contents": "read", "pull_requests": "write"}},
-		},
-		Authorization: compiler.PlanAuthorization{GitHubTokenActions: []string{"owner/action@v1"}},
-	}
-	wantMessage := `Job "build" needs GITHUB_TOKEN, but job-level permissions are unsupported for hosted GITHUB_TOKEN issuance. Move the permissions map to the workflow top level.`
-	wantDetail := `Cause: action "owner/action@v1" defaults an input to github.token. Effective permissions: contents: read, pull-requests: write.`
-	message, detail := githubTokenAdmissionDiagnostic(artifact, "GitHub workflow access tokens do not support job-level permissions")
-	if message != wantMessage || detail != wantDetail {
-		t.Fatalf("githubTokenAdmissionDiagnostic() = %q, %q", message, detail)
-	}
-}
-
-func TestUnprivilegedUploadAllowsPublicAndDockerfileActionCapabilities(t *testing.T) {
-	for _, capabilities := range [][]string{nil, {"network"}, {"docker"}, {"docker", "network"}} {
-		bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-			Workflow:             plan.Workflow{LogicalJobID: "action-job"},
-			RequiredCapabilities: capabilities,
-			Steps:                []plan.Step{{ID: "action", Kind: "uses", Uses: "owner/example@commit"}},
-		}, Authorization: compiler.PlanAuthorization{DockerCapabilitySources: []string{"dockerfile-actions"}}}}}
-		if err := validateUnprivilegedBundle(bundle); err != nil {
-			t.Fatalf("validateUnprivilegedBundle(%v) error = %v", capabilities, err)
-		}
-	}
-}
-
-func TestUnprivilegedUploadRejectsDockerWithoutCompilerProvenance(t *testing.T) {
-	bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-		Workflow:             plan.Workflow{LogicalJobID: "unproven-docker"},
-		RequiredCapabilities: []string{"docker"},
-	}}}}
-	err := validateUnprivilegedBundle(bundle)
-	var finding *compiler.ProcessingFinding
-	want := `Job "unproven-docker" requires Docker without matching compiler provenance. Hosted runs support only verified Dockerfile actions and bounded job or service containers.`
-	if err == nil || !errors.As(err, &finding) || finding.Message != want || finding.Detail != "" {
-		t.Fatalf("validateUnprivilegedBundle() error = %v, want Docker provenance rejection", err)
-	}
-}
-
-func TestUnprivilegedUploadAdmitsCompilerVerifiedContainerProvenance(t *testing.T) {
-	for _, test := range []struct {
-		name    string
-		job     plan.Job
-		sources []string
-	}{
-		{name: "job", job: plan.Job{Container: &plan.Container{Image: "node:24"}}, sources: []string{"job-containers"}},
-		{name: "service", job: plan.Job{Services: map[string]plan.ServiceContainer{"redis": {Image: "redis:7"}}}, sources: []string{"service-containers"}},
-		{name: "dynamic services", job: plan.Job{ServicesExpression: "${{ fromJSON(needs.build.outputs.services) }}"}, sources: []string{"service-containers"}},
-		{name: "all", job: plan.Job{Container: &plan.Container{Image: "node:24"}, Services: map[string]plan.ServiceContainer{"redis": {Image: "redis:7"}}}, sources: []string{"dockerfile-actions", "job-containers", "service-containers"}},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-				Workflow: plan.Workflow{LogicalJobID: "container-job"}, RequiredCapabilities: []string{"docker", "network"}, Container: test.job.Container, Services: test.job.Services, ServicesExpression: test.job.ServicesExpression,
-			}, Authorization: compiler.PlanAuthorization{DockerCapabilitySources: test.sources}}}}
-			if err := validateUnprivilegedBundle(bundle); err != nil {
-				t.Fatalf("validateUnprivilegedBundle(%v) error = %v", test.sources, err)
-			}
-		})
-	}
-}
-
-func TestUnprivilegedUploadRejectsMismatchedContainerProvenance(t *testing.T) {
-	for _, test := range []struct {
-		name    string
-		job     plan.Job
-		sources []string
-	}{
-		{name: "claim without container", sources: []string{"job-containers"}},
-		{name: "container without claim", job: plan.Job{Container: &plan.Container{Image: "node:24"}}},
-		{name: "unknown source", sources: []string{"docker-plugin"}},
-		{name: "unsorted source", job: plan.Job{Container: &plan.Container{Image: "node:24"}}, sources: []string{"job-containers", "dockerfile-actions"}},
-	} {
-		t.Run(test.name, func(t *testing.T) {
-			bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-				Workflow: plan.Workflow{LogicalJobID: "container-job"}, RequiredCapabilities: []string{"docker", "network"}, Container: test.job.Container,
-			}, Authorization: compiler.PlanAuthorization{DockerCapabilitySources: test.sources}}}}
-			if err := validateUnprivilegedBundle(bundle); err == nil || !strings.Contains(err.Error(), "unsupported Docker access") {
-				t.Fatalf("validateUnprivilegedBundle(%v) error = %v", test.sources, err)
-			}
-		})
-	}
-}
-
-func TestUnprivilegedUploadRejectsKnownGitHubServiceActions(t *testing.T) {
-	tests := []struct {
-		action  plan.ActionLock
-		service string
-	}{
-		{plan.ActionLock{Source: "github", Repository: "actions/upload-artifact", Path: "merge"}, "artifact"},
-	}
-	for _, test := range tests {
-		name := test.action.Repository
-		if test.action.Path != "" {
-			name += "/" + test.action.Path
-		}
-		t.Run(name, func(t *testing.T) {
-			bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-				Workflow: plan.Workflow{LogicalJobID: "service-action"},
-				Actions:  []plan.ActionLock{test.action},
-			}}}}
-			err := validateUnprivilegedBundle(bundle)
-			var finding *compiler.ProcessingFinding
-			want := `Action "actions/upload-artifact/merge" requires the GitHub Actions artifact service, which hosted runs do not provide. Replace the action or use a runtime profile that provides this service.`
-			if err == nil || !errors.As(err, &finding) || finding.Message != want || finding.Detail != "" {
-				t.Fatalf("validateUnprivilegedBundle(%#v) error = %v", test.action, err)
-			}
-		})
-	}
-}
-
-func TestUnprivilegedUploadAllowsOnlyAuditedCacheCommits(t *testing.T) {
-	for _, commit := range actionintegration.CacheCommits() {
-		for _, path := range []string{"", "restore", "save"} {
-			action := plan.ActionLock{Source: "github", Repository: "actions/cache", Path: path, Commit: commit}
-			bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-				Workflow: plan.Workflow{LogicalJobID: "cache"},
-				Actions:  []plan.ActionLock{action},
-			}}}}
-			if err := validateUnprivilegedBundle(bundle); err != nil {
-				t.Fatalf("validateUnprivilegedBundle(%#v) error = %v", action, err)
-			}
-		}
-	}
-
-	action := plan.ActionLock{Source: "github", Repository: "actions/cache", RequestedRef: "v6.0.2", Commit: strings.Repeat("0", 40)}
-	bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-		Workflow: plan.Workflow{LogicalJobID: "cache-old"},
-		Actions:  []plan.ActionLock{action},
-	}}}}
-	err := validateUnprivilegedBundle(bundle)
-	if err == nil || !strings.Contains(err.Error(), "v6.1.0") {
-		t.Fatalf("validateUnprivilegedBundle(%#v) error = %v", action, err)
-	}
-	var finding *compiler.ProcessingFinding
-	if !errors.As(err, &finding) || finding.Message != "actions/cache v6.0.2 is unsupported. Use a supported version from https://github.com/buildkite/buildkite-gha/blob/main/docs/compatibility.md#cache-action." || !strings.Contains(finding.Detail, "Buildkite cache-v2 service") || !strings.Contains(finding.Detail, action.Commit) || strings.Contains(finding.Message, action.Commit) {
-		t.Fatalf("hosted cache diagnostic = %#v", finding)
-	}
-}
-
-func TestCacheServiceRequiredUsesOnlyAuditedCacheLocks(t *testing.T) {
-	required, err := cacheServiceRequired([]plan.ActionLock{{Source: "github", Repository: "owner/action", Commit: strings.Repeat("a", 40)}})
-	if err != nil || required {
-		t.Fatalf("ordinary action cache requirement = %v, %v", required, err)
-	}
-
-	locks := []plan.ActionLock{{Source: "github", Repository: "owner/action", Commit: strings.Repeat("a", 40)}}
-	for _, commit := range actionintegration.CacheCommits() {
-		locks = append(locks[:1], plan.ActionLock{Source: "github", Repository: "actions/cache", Path: "restore", Commit: commit})
-		required, err = cacheServiceRequired(locks)
-		if err != nil || !required {
-			t.Fatalf("cache requirement for %s = %v, %v", commit, required, err)
-		}
-	}
-
-	locks[1].Commit = strings.Repeat("b", 40)
-	if _, err := cacheServiceRequired(locks); err == nil || !strings.Contains(err.Error(), actionintegration.CacheCommit) {
-		t.Fatalf("unsupported cache lock error = %v", err)
-	}
-}
-
-func TestUnprivilegedUploadAllowsNativeDownloadArtifactAdapter(t *testing.T) {
-	bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{Workflow: plan.Workflow{LogicalJobID: "consumer"}, Actions: []plan.ActionLock{{Source: "github", Repository: "actions/download-artifact", Commit: actionintegration.DownloadArtifactCommit}}}}}}
-	if err := validateUnprivilegedBundle(bundle); err != nil {
-		t.Fatal(err)
-	}
-}
-
-func TestUnprivilegedUploadAllowsNativeUploadArtifactAdapter(t *testing.T) {
-	bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-		Workflow: plan.Workflow{LogicalJobID: "artifact-producer"},
-		Actions: []plan.ActionLock{{
-			Source: "github", Repository: "actions/upload-artifact",
-			Commit: actionintegration.UploadArtifactCommit,
-		}},
-	}}}}
-	if err := validateUnprivilegedBundle(bundle); err != nil {
-		t.Fatalf("validateUnprivilegedBundle() error = %v", err)
-	}
-}
-
-func TestUnprivilegedUploadDoesNotBroadenKnownServiceActionIdentity(t *testing.T) {
-	for _, action := range []plan.ActionLock{
-		{Source: "workspace", Repository: "actions/cache"},
-		{Source: "github", Repository: "actions/cache", Path: "nested"},
-		{Source: "github", Repository: "actions/upload-artifact", Path: "nested"},
-		{Source: "github", Repository: "actions/download-artifact", Path: "nested"},
-		{Source: "github", Repository: "owner/action"},
-	} {
-		bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-			Workflow: plan.Workflow{LogicalJobID: "ordinary-action"},
-			Actions:  []plan.ActionLock{action},
-		}}}}
-		if err := validateUnprivilegedBundle(bundle); err != nil {
-			t.Fatalf("validateUnprivilegedBundle(%#v) error = %v", action, err)
-		}
-	}
-}
-
-func TestBundleUsesActionsDetectsStepsAndLocks(t *testing.T) {
-	for _, job := range []plan.Job{
-		{Steps: []plan.Step{{Kind: "uses"}}},
-		{Actions: []plan.ActionLock{{ID: "a-deadbeefdeadbeef"}}},
-	} {
-		if !bundleUsesActions(compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: job}}}) {
-			t.Fatalf("bundleUsesActions() = false for %#v", job)
-		}
-	}
-	if bundleUsesActions(compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{Steps: []plan.Step{{Kind: "run"}}}}}}) {
-		t.Fatal("bundleUsesActions() = true for shell-only plan")
-	}
-}
-
-func TestUnprivilegedUploadAllowsCapabilityFreeConcurrentShellSteps(t *testing.T) {
-	bundle := compiler.Bundle{Plans: []compiler.PlanArtifact{{Job: plan.Job{
-		Workflow: plan.Workflow{LogicalJobID: "concurrent-job"},
-		Steps: []plan.Step{
-			{ID: "background", Kind: "run", Command: "true", Background: true},
-			{ID: "wait", Kind: "wait", Targets: []string{"background"}},
-			{ID: "wait-all", Kind: "wait-all"},
-			{ID: "cancel", Kind: "cancel", Targets: []string{"background"}},
-		},
-	}}}}
-	if err := validateUnprivilegedBundle(bundle); err != nil {
-		t.Fatalf("validateUnprivilegedBundle() error = %v", err)
-	}
-}
-
 type cliCommand struct {
 	dir   string
 	name  string
@@ -6004,18 +4840,6 @@ func TestUploadArgsParsesPlatformRuntimeDistributions(t *testing.T) {
 		if _, err := parseUploadArgs(test.args); err == nil || !strings.Contains(err.Error(), test.want) {
 			t.Fatalf("parseUploadArgs(%#v) error = %v, want %q", test.args, err, test.want)
 		}
-	}
-}
-
-func TestValidateActionCacheRequiresHostedProfile(t *testing.T) {
-	var stdout, stderr bytes.Buffer
-	if code := Run([]string{"validate", "--action-cache-dir", t.TempDir(), "workflow.yml"}, &stdout, &stderr, "dev"); code != 2 || !strings.Contains(stderr.String(), "requires --profile hosted") {
-		t.Fatalf("Run() code = %d, stderr = %q", code, stderr.String())
-	}
-	stderr.Reset()
-	missing := filepath.Join(t.TempDir(), "missing")
-	if code := Run([]string{"validate", "--profile", "hosted", "--event", "push", "--action-cache-dir", missing, "workflow.yml"}, &stdout, &stderr, "dev"); code != 2 || !strings.Contains(stderr.String(), "cache root is not a non-symlink directory") {
-		t.Fatalf("Run() code = %d, stderr = %q", code, stderr.String())
 	}
 }
